[roll] Roll fuchsia [archivist] GC stale ComponentLogStats
Cobalt often fails to receive accurate ComponentLogStats when a
component crashes restarts because when it disconnects the last clone of
the Arc<Mutex<ComponentLogStats>> gets dropped and a new object is
lazily instantiated if/when the component restarts.

This change introduces a timer that reviews Arc<ComponentLogStats> for
freshness and drops references to stale ones. This ensures a lower bound
on how long a Arc<ComponentLogStats> is retained, regardless of whether
the number of clients retaining them drops to 0 (as in the case with a
component crash or restart). The timer is currently set to trigger every
hour. For reference, currently Cobalt collects these metrics every 15
minutes.

BUG=56527
TEST=fx test //src/diagnostics/archivist:archivist_tests -- logs::stats::tests::component_stats_retained_then_dropped

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/428414
Original-Revision: b24db49f866c557db9d6ec716ed9226e802d7795
GitOrigin-RevId: dc1be38c6fcdb7beebdcd1648c34cea931ad2cd4
Change-Id: I6d4dbf6454ca13a08a54b9a2f2a7c6df74605581