[roll] Roll infra/recipes [fuchsia/build] set got_revision to baseline integration rev
Previously the got_revision was set to the baseline of whatever repo
triggered the build. For the binary size diffs, we always want the
baseline to come from integration CI, so we need got_revision to be a
revision from the integration repo.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/429359
Original-Revision: 2af3e9d0b8ee05bc4aaa9902a28e3a18386b2dc4
GitOrigin-RevId: bba66c4c561dc2d0a68bb31d7c3319caf5771060
Change-Id: I70b1a0eb186803491e92cf9884dbefccdcf5c8ad