[roll] Roll fuchsia [fuzz] Resolve SSH credentials from manifest
The location of the default SSH credentials used in connecting to a
Fuchsia device has changed. This change updates fuzz to use a manifest
to find the appropriate path to the default SSH private key (identity)
and authorized_keys.

The aforementioned manifest is generated by a jiri hook on update and is
not expected to be absent in any Fuchsia tree unless the user has
removed it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/419481
Original-Revision: 4ca07f9a7e492669a78f5c143e106fe9306b34d5
GitOrigin-RevId: 2eaa20e42f2532b8fdc620f953a063e46a93dc85
Change-Id: I3366a1228cdad53e4f3117e53cdafac556ae3f35