[roll] Roll fuchsia [build] Restore zircon toolchain commands equivalence.
A recent change [1] moved the @ symbol used for response files
in quotes, but forgot to update the zircon_toolchain.gni file
as well. As a consequence, the build was not broken, but the
compare_toolchains.py script now complains that the commands
are not equivalent anymore between the Zircon and the Fuchsia
build.

This CL trivially fixes the issue.

Original-Bug: 60296
Testing: The command [2] will succeed

[1] https://fuchsia-review.googlesource.com/c/fuchsia/+/429174
[2] zircon/public/canaries/compare_toolchains.py --clean /tmp/results

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/429614
Original-Revision: 860f2a37baedc4520f1a431cf69de0dda7613829
GitOrigin-RevId: fbb2ee251bc4c39de3cb343f6ff5fef8dea40f39
Change-Id: I60a4881cec94668d09e47dc44914fb556a2458bf