[roll] Roll fuchsia [fshost] Fix use-after-free bug.
ASan stack-user-after-return detected a bug in this codebase.
Specifically, a local VMO object handle was passed by reference
to a spawned thread. This CL changes that so that a copy of this object
is passed.

The ASan stacktrace can be found here:
https://pastebin.pl/view/82790e2a

Original-Bug: 59214

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/429434
Original-Revision: 999693ce6c255882532813039f5ef3d966d6bbbc
GitOrigin-RevId: 885c2fb8c7af5772cd96e52a9232b0dba7b91110
Change-Id: I46ddb1fc7188cbc48c29d0e22b9b6a99b9efcb58