[roll] Roll infra/recipes [fuchsia_roller] Never roll backwards
We've recently noticed that the fuchsia-roller sometimes rolls
"backwards" - i.e. to an older commit. This happens when gitiles returns
an older revision for the current master branch than the revision that's
currently pinned in integration. It's not clear why this happens, but
it's easy enough to avoid rolling backwards in this case by simply
failing the roller rather than rolling to an older commit (commit A is
considered older than commit B if their hashes are unequal and `gitiles
log A..B` is empty).

Original-Bug: 60361
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/429364
Original-Revision: 74d59fcb7a2d7b346fc653366d48e600ea9c7e97
GitOrigin-RevId: 1280ab99b40c11764a58ba749230d0da616013fc
Change-Id: I1b9b50ab173eff8d6297ca40e2f4f0ab8182f636