[roll] Roll infra/recipes [fuchsia] Use env_name for resultdb upload
`spec.test.device_type` is only set for non-sharded builders, so it
can't always be used as a resultdb tag. A better candidate is the "test
environment" (e.g. "AEMU-x64"), which we already use for tagging test
results in Sponge. So I replaced the `device_type_for_test` resultdb tag
with a `test_environment` tag containing the test environment name.

Unfortunately, the only way to propagate that value to the orchestrator
without adding additional metadata files to the orchestration inputs is
to read it from the each shard's Swarming task request tags. This
is slightly fragile and a little awkward, especially because the tags
are a list with items of the form "key:value" rather than a dict, so we
have to do some parsing to retrieve the desired value.

The `env_name` property of the `FuchsiaTestResults` class now contains
the testing environment name. Previously it contained the shard name,
making it a bit of a misnomer, because the shard name also contains the
shard index and thus differs between shards that use the same
environment. So I renamed the shard name property to the more
appropriate `_shard_name`.

I also clarified in fuchsia.proto that `device_type` is only valid for
non-sharded builders.

Original-Bug: 60001
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/428896
Original-Revision: a54e97f9887ec6b9ee17d644e311e1ccebae2e90
GitOrigin-RevId: 4c25a46c7b1e68a7fcf62559b872470df0bf678c
Change-Id: I958745a36421fc27673cf8f01ae1ce307c32b445