[roll] Roll fuchsia [libc] Don't use -ffreestanding
The libc build always used -ffreestanding since that's what
upstream musl did originally.  This is theoretically necessary to
avoid the compiler e.g. recognizing the implementation of a
standard string function and replacing it with an outcall to
itself.  But libc today doesn't seem to have any actual problems
of that sort in practice, and this prevents good optimizations.
Dropping this switch reduces libc.so code size a bit and almost
certainly performs better as well.

In ASan builds, -ffreestanding was having the additional effect
that memcpy and memset calls, both those generated by the
compiler and explicit calls, were not being rewritten to
__asan_memcpy and __asan_memset.  That caused an unintended lack
of some checking across libc code.  But it also implicitly
avoided a sticky bootstrapping problem.  Calls to __asan_memcpy
and __asan_memset are PLT calls, which cannot be used before
dynamic linking has resolved PLTs.  Once PLTs are resolved, these
functions are compiled with ShadowCallStack by default and so
require the stack setup before they can be used.  So these both
must be avoided in the dynamic linking code and in the early
startup code before secondary stacks are set up.  We had an
existing trick for this, which had come up before in pre-PLT uses
of memset.  After dropping -ffreestanding it came up in pre-PLT
uses of memcpy, and also post-PLT but pre-stack-setup uses of
both, so we extend the same trick to memcpy and apply it in the
early setup code as well as the dynamic linking code.

Original-Bug: 60123
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/428524
Original-Revision: 110fc221bf5a5b6dea7cab66501898690dde8ad7
GitOrigin-RevId: 9e0a6a04d84018e217eb6d981aa63179a8ae3b02
Change-Id: I3214127c4f362f0a20c71c1e75619438a31dc52b