[roll] Roll fuchsia [fidl][llcpp] Prevent overflowing unknown handle buffer
This asserts that we don't overflow the unknown handle buffer. A full
solution would close handles that overflow the buffer, but this adds
complexity - particularly so because there are different handle closing
implementations for handles and handle info.

Since this is a case that shouldn't occur ever currently, I favored
keeping the implementation simple and checking that it doesn't happen
with a ZX_ASSERT.

Original-Fixed: 59999
Test: There are no tests because it will crash when this condition is
met and it is difficult to construct a test that can detect this
condition.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/429636
Original-Revision: e5c5099b2ba70d8a13643b49c9725f3bc048d49b
GitOrigin-RevId: b2f7c50f24ed7d5356acfb951df5f468e8b7b2a8
Change-Id: I248d1d2e0345e72c1577525e23905f7cf914ef13