[roll] Roll fuchsia [flatland] Inject DisplayController into Renderer
Images used as textures need to have the buffer collections
they are associated with imported to the display so that
the display can recognize the images when we offline rendering
data to it.

This importing needs to happen at the same time as when we
set the Vulkan constraints on the renderer itself, so this CL
injects the display controller into the renderer so it can
control buffer collection management for the DC as well. This
also allows the DC buffer collection registration to happen
inside the flatland class without exposing the DC directly to
Flatland.

Lastly, this CL makes it so that the renderer and the display
share the same unique IDs for buffer collections, in order to
forego the need to keep track of mappings between them. To do
this, a global function has been added to display/utils.h to
generate ids atomically.

Testing - Unit tests in display_unittest.cc and engine_unittest.cc
updated to account for the change.

MULTIPLY: flatland_renderer_unittests (fuchsia): 30

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/429458
Original-Revision: cf79f817a93d8b73e1f5660417056d1b90d826df
GitOrigin-RevId: 4a9d2f51bddfd88f9bb8efe849d9601e03fe392f
Change-Id: I790541d2d8f3b11e68738a84bd56280f879b4339