[roll] Roll fuchsia [brcmfmac] Do not log initial MAC address upon iface creation
The initial MAC address of an iface immediately after the driver
creates it is not meaningful. It is always reassigned after
creation. Leaving the initial MAC address in the syslog can be
misleading as the MAC address can be completely different than what it
will eventually become. For example, the initial MAC address of the
client iface is always 00:00:00:00:00:00 before its reassigned.

Test: None necessary. Logging change only.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430338
Original-Revision: a1bcb38e7e632457d703eb29657635962f30fe1e
GitOrigin-RevId: dafe999543e8dd342f50775e24e4f0963571e868
Change-Id: Ia5380a3e6a0d29fd1f10429ec828e6db5f4d284c