[roll] Roll fuchsia [soundplayer] teach example util to play sounds from /tmp
This CL adds a feature to soundplayer_example that has it play all
the sounds in its isolated /tmp, if there are any such files. When
such files are found, the hard-coded sequence is not played.

This feature is useful for establishing that sounds will play
correctly with soundplayer.

Here's a host-side bash function that can be used to play a directory
full of files:

function playmanysounds() {
  if (($# != 1)); then
    echo "usage: playmanysounds <directory>"
    return 1
  fi
  fx shell mkdir /tmp/r/sys/fuchsia.com:soundplayer_example:0#meta:soundplayer_example.cmx
  for file in $1/*; do
    echo copying $file
    fx cp --to-target $file /tmp/r/sys/fuchsia.com:soundplayer_example:0#meta:soundplayer_example.cmx
  done
  fx shell run soundplayer_example
  fx shell rm -rf /tmp/r/sys/fuchsia.com:soundplayer_example:0#meta:soundplayer_example.cmx
}

Test: example...no automated tests
      The bash function above can be used for manual test.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430350
Original-Revision: c8fad40cf114ae14a122e5ac456a7ffedda6b79b
GitOrigin-RevId: c337a8438470c2a12dd35e209b11bbdbd5e53f97
Change-Id: Ide88391d10b4c6e87affff79e10475c1f8ebb8ca