[roll] Roll fuchsia [iwlwifi] Enable driver but exclude from build
Instead of writing bind rules that always fail, just exclude the driver
from the build. Only the driver module itself is excluded, so the tests
that depend on the driver source will still be compiled and tested in
continuous integration. The only effect is that the driver manager will
not read the driver's info at runtime, or execute its bind rules, unless
the driver module is explicitly added to a build. And the driver won't
be compiled for most developers building a core or product build without
all the tests.

To run the driver, add the following to fx set:
--with-base //garnet/packages/prod:wifi_intel

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430057
Original-Revision: b43d604c7be32741459fd067620ccdf867b442f4
GitOrigin-RevId: 88a47e35e42e7804951e608694ac66fa7860c5a3
Change-Id: Ie244287a9349e10c5f726b93e21699435d068588