[roll] Roll fuchsia [build] Require test_durations_file gn arg
Instead of falling back to the default test durations file if the gn arg
is unset, pass an empty test durations file to testsharder. This empty
test durations file will likely cause testsharder to fail because it
doesn't contain a default entry, but that's ok because we'll make sure
to always have the recipe set `test_durations_file` for builders that
use testsharder.

This is necessary because we'll soon be consuming test durations from
CIPD packages. At that point the duration files will no longer be
located in the same directory regardless of whether the checkout used
the public or internal manifest. So the decision of which default
duration file to fall back to in case the builder-specific duration file
doesn't exist must be made by the recipes.

This should not land until after the corresponding recipe change lands;
otherwise any new builders that don't yet have duration files would
break.

Original-Bug: 46010
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430256
Original-Revision: 0f983f71355b2476272edbbae375ab4bc52738a3
GitOrigin-RevId: b097dd3bab7303a1456e33ef1c1a1311db21b768
Change-Id: I6887910bcfa0e6a31424cdeda13de8d51160310c