[roll] Roll fuchsia [flatland] Fix GraphLink and ContentLink status APIs.
The FIDL documentation for GraphLinkStatus states "GraphLinks will
be informed when they are actively attached to an output display or
not". Prior to this change, LinkSystem::UpdateLinks() would indicate
that the child was attached to the display if the parent Flatland's
link-associated handle is in the live set, when in fact it should
be when the child Flatland's link-associated handle (known as the
link_origin).

Additionally, the ContentLinkStatus states that CONTENT_HAS_PRESENTED
will not fire until the child Flatland's content "has been rendered
to an output device", when in fact it should fire as soon as the
child has called Present() at least once (indicating that it has
content ready to render).

Also removes unneccesary computation from ProcessMainLoop() so that
it only performs link-related logic and can be renamed to
UpdateLinks() and adds clarificationc comments that
LinkSystem::UpdateLinks() is unit tested in the Flatland unit tests.

TEST=fx test -o flatland_unittests
MULTIPLY: flatland_unittests (fuchsia): 30

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/428048
Original-Revision: c8f0db6733567ff671d4b784f939000ed5d65433
GitOrigin-RevId: 444cbf4fc137a99fa4c3cbf11527bdf9e2514c1d
Change-Id: Idbc3c3cab46fdae782531fd1b5c3a42077392b10