[roll] Roll fuchsia [bootserver] Pass in authorized keys to Boot().
The bootserver tool takes in an authorized keys file as a flag, but I
incorrectly parsed that as a private key file because botanist used a
private key file and derived the authorized keys from it and I just
copied the logic. This change calls for the authorized_keys bytes to get
passed to Boot() instead of the private SSH signers so that the
bootserver tool can just pass its authorized_keys as is and botanist
will have to convert first before passing the keys to Boot().

Test: tested by CQ

Original-Bug: 38517
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430540
Original-Revision: fc712d47a209152f7aef79ed07f2c745c1f7d124
GitOrigin-RevId: d82ef6c0318e8dd0753f621ccfdcc086cb3ff5fd
Change-Id: I1d2b1e62aa563911f4935cb3cbc3dd2151a4f970