[roll] Roll fuchsia [ui-input] refactor input-synthesis tests
Refactor the unit test code, to separate the collection
of events from the validation of those events. This enables
non-exact-comparison, which we'll need for a timestamp
test which we'll be adding shortly.

Along the way: remove the use of spawn(), in favor of
having the collection-and-validation-task run via
the run_singlethreaded() attribute macro.

Original-Bug: 60438
Test: fx test input-synthesis-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430034
Original-Revision: 98030b09bbf4e54553ba1073c5417de94edfefdb
GitOrigin-RevId: 3e505a424dc12070581e5a9945f1b9347bfdb906
Change-Id: I086ddea9f2ebb74998c0cf5ab4322ada59162673