[roll] Roll fuchsia [fidl][rust] Add support for flexible enums
This CL adds support for flexible enums in the FIDL Rust bindings.
Unlike strict enums, they have an extra variant __Unknown which is
returned when the value on the wire is not recognized. Here is the API:

    from_primitive(prim: $prim_ty) -> Option<Self>  // same as strict
    into_primitive(self) -> $prim_ty                // same as strict
    unknown() -> Self                               // flexible only
    from_primitive_flexible(prim: $prim_ty) -> Self // flexible only
    validate(self) -> Result<Self, $prim_ty>        // flexible only
    is_unknown(&self) -> bool                       // flexible only

In addition, a flexible enum `Foo` gets a macro `FooUnknown!()` to be
used in match expressions. This expands to `_` normally, but in the
future there will be a build mode where it expands to the unknown
variant, allowing the detection of missing cases.

The __Unknown variant is marked deprecated to discourage using it
directly. If a user references Foo::__Unknown, they get this warning:

    WARNING: Use Foo::unknown() to construct, FooUnknown!() to exhaustively match

Test: fx test //src/lib/fidl/rust
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/428515
Original-Revision: fbc55ac22e5d9d86195ed2376c9dd2d677994296
GitOrigin-RevId: b9dc452742a5b629dc457b4c525eb7c1899b5752
Change-Id: I581e52629e358bab980235321cbd818807c225f8