[roll] Roll fuchsia [drivers][time] Propagate clock.backstop boot arg to drivers
The RTC driver needs the same backstop as the rest of the system
in order for the kernel UTC time to stay in sync with the userspace
UTC clock. This is only necessary until UTC is removed from the kernel.

Original-Bug:60452

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430794
Original-Revision: 48b70e16e521b79570f86e5ed6339f450ba30b2e
GitOrigin-RevId: c3aa4c1cb7d9c3898e658dce58f8fba8e56580ca
Change-Id: I8fc12b5be058c8ed13db6ae1eac0c2e1a9a682dc