[roll] Roll fuchsia [bootserver] Only get images that exist.
When trying to get images with a local manifest, we assumed all images
existed and failed if it didn't. However, not all images are guaranteed
to exist and although we skip non-existent images from GCS, we didn't
have that check for the local images. This adds that check.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430004
Original-Revision: 852a6d335827f0f27b31fd2b31f63f74e441e266
GitOrigin-RevId: 9574eabc5f16ac08656fd41eb0856ee22202d5e4
Change-Id: I16da1ca5864a94ebb16f5c4547f6ad392573d2bf