[roll] Roll fuchsia [audio] Rename TDM metadata to better align to codec API
We rename PCM to TDM1 to match the codec API since much
of the aml-g12-tdm is now parameterized allowing more freedom
in the configuration similar to the freedom in TDM1.
We also add Stereo to the LeftJustified name.

Test: Build only, refactoring.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430006
Original-Revision: 0af8637e89d629f592e6ab4b41ab0abb861ad681
GitOrigin-RevId: 8dcd200c8a949298cb0762019bb14fb0688e2f5c
Change-Id: I70493c0d6cae2b1eff935a8d665dc95295a6b935