[roll] Roll fuchsia [audio][codecs] Remove self initialization requirement
Since there are timing requirements tied to the TDM bus
out of the control of some codec drivers we remove the
auto initialization requirement such that a manual codec
Reset call is used to reset and initialize. We remove
auto initialization of drivers that performed it.

Test: tas58xx-test
Test: max98373-test
Test: tas5782-test
Test: tas27xx-test
Test: Manual audio-driver-ctl check on Astro

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/428832
Original-Revision: ccd8b8f09516e7076f5070fe1ffed53b8adcb376
GitOrigin-RevId: f88535e9ec130c8b85583da42aa1eb50df2c104a
Change-Id: I61a9e3f039619deda98591aff638d7898c604847