[roll] Roll fuchsia [build] Fix `fx format-code` and `fx lint` for FIDL files.
The 'fidl-format' and `fidl-lint` tools were recently moved to
the Fuchsia build, so update the scripts to look for the binaries in
the right place (and rebuild them if necessary).

For testing, do the following:

  - Modify a liberal .fidl file (e.g. messing with indentation).
  - Do "fx clean"
  - Run "fx format-code"
  - Run "fx lint"

Due to the clean build, this will rebuild the fidl-format and
fidl-lint tools respectively, then processing the .fidl file
with them. Then verify manually that it was fixed, since the
tools don't print anything in that case.

Original-Bug: 49742
Testing: See above
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430875
Original-Revision: bbb0c49d79f4a979e5e401b7694ae7ec4423fb14
GitOrigin-RevId: caf90cfc7832268a5a554b27d36c0c6a592453e2
Change-Id: Idbd95893cdd426675b277f17943a1b2a543daf71