[roll] Roll fuchsia [cts] use prebuilt python to run test
Introduce python_host_test template to make this easy for other tests
to re-use. This results in tests that are totally self-contained, rather
than relying on python3.8 being in the PATH when they happen to get
executed. This should fix the attached issue, which is a result of tests
relying on the PATH.

Also add cts/build:tests to cts:cts so that the tests get run in CI,
which made it much easier to test this change.

Original-Fixed: 60260
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430344
Original-Revision: a01e29b9e5426fac3f9d2cda9d4445413a19dfad
GitOrigin-RevId: 005bd88cb837e116ecd250358b0d6fcfb13d92ce
Change-Id: I0381c18ecf971343b775e7d4e5c061785715feed