[roll] Roll fuchsia [kernel][sched] Add ThreadQueueEvents to tracing
Add trace probes to the scheduler which record when a thread enters or
leaves a CPU run queue, or when it becomes active on that CPU.
Also, record the effective affinity mask of the thread at this point
in time.

This allows for an alternate method of computing utilization
statistics, one which can take into account the affinity restrictions
present in the system (simply using Run-Q lengths cannot do this).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/418722
Original-Revision: 41fec0f9fad392c3b0e7a15b457255ba3d0b67d7
GitOrigin-RevId: a8d912b5acbe6a02820b21104b48cb795e8722f4
Change-Id: I056d79095c4331e5bd04c01c93e71b844c9ed080