[roll] Roll fuchsia [netstack] Observe v6 LL addr in if watcher test
Fix a flake in netstack_integration_test::test_interfaces_watcher
where the configuration of an IPv6 link-local address after the
interface is brought up may be observed and cause the assertion of
events to fail. After the interface is brought up, wait until the
interface up and the LL addresses are observed. After adding the IPv4
address, check that the new address set is equal to the LL addresses
observed previously plus the added address. The reverse of these
assertions are made when the IPv4 address is removed and when the
interface is brought down (which causes exactly one of the LL
addresses to be removed).

Includes some refactoring to improve code quality.

Test: fx test netstack_integration_test -- \
          --exact fidl::test_interfaces_watcher
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/429887
Original-Revision: 61c8902534e6464d4fd753e9fe9f2b2ee05c23e8
GitOrigin-RevId: 079b190dc35c18adde2fbfcd7b62455939abc384
Change-Id: Iec4954cbf68aba877f5d070f8a74e74c7adec349