[roll] Roll fuchsia [modular][testing] Ensure test harness is torn down after test fixtures
TestHarnessFixture will now stop the test harness in TearDown, before
the fixture is destructed.

This ensures that agents and shells created in tests and stored in
fixture fields are destructed only after the harness is stopped.
In particular, this is needed to ensure that sessionmgr is connected
to the SessionShell protocol through its teardown process, as it may
need to call DetachView when deleting stories.

Original-Bug: 60191
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430957
Original-Revision: be7e4e03787a4f6b6fe3665b892207058be4f85b
GitOrigin-RevId: 9134fe6ddcd17a509a949993633f3a60062737a1
Change-Id: I33e72c61eef570822bac35dd40c0c255159a4e8e