[roll] Roll fuchsia [libc] Replace strcmp call in start_main
There's a call to strcmp before we've finished initializing the
sanitizer runtime. However, this might crash, because strcmp might
itself be instrumented.

This isn't a problem for sanitizers like asan, because they fail open
when the shadow memory is zero initialized. But for tsan, the
instrumentation functions expect a TLS pointer variable to be
initialized, which causes a null pointer dereference crash in
__tsan_func_entry.

Original-Bug: 45118
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/429879
Original-Revision: f00849ad2e2cf4be765ef6f49743f13251cb18d1
GitOrigin-RevId: 7a3587b3a936fa22a73bacdaae79cc195eb4298e
Change-Id: I699cfbbc1df95765e18749c366eb784f86ac7349