[roll] Roll fuchsia [cts] Do CTS Dep Verification In a Python Script
Dependency verification is too much work to be done during generation
time, so it has to be done during build time. The Python script searches
for dependencies in the allowed_cts_deps list and looks for a CTS file
if the dependency isn't there. CTS files are written at generation time,
and are assumed to be correct until the build checks the dependencies.

Renames the cts_element targets because it's used as part of the output
of the inner action now.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/419484
Original-Revision: 2e2a589850755cf2d439da800a525331d5eeeb5f
GitOrigin-RevId: a0bf80d55a2a15e960fa2a253e2845a10d852b22
Change-Id: I706eb7377d76c0916efd3fa58a8e99d0909e591b