[roll] Roll fuchsia [fdf] Remove <zircon/driver/binding.h>
Move the contents into the DDK and replace all references to the old
header with that. This is both a clean-up, there's no need for this
header to be in the zircon public includes, and will simplify future
changes to how drivers use bind rules.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430319
Original-Revision: bcf4e716764fd102d886c82a0ac53185a0b53e5b
GitOrigin-RevId: d5637a5b3aabd931008b7c55138f6ab55721a35a
Change-Id: Ia250cb1551638d3737bcdedd9f4e0d34e116dece