[roll] Roll fuchsia [fatfs] Fix cluster not zeroed bug.
In a different CL, I changed the fs_tests so the disk is filled with
non-zero data which highlighted this bug: any clusters allocated to
the root directory would not get zeroed.

Fatfs relies on File::is_dir() to determine whether or not to zero a
cluster when allocating a new one. Previously, is_dir() would return
false for the root directory, which would cause extra root directory
clusters to not be zeroed. Fix this so that the clusters are correctly
zeroed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/431959
Original-Revision: 96db90c176ac2f593f24952242fa2a4c2f3bcb27
GitOrigin-RevId: 2d506737030301e288b93c27418634b637c82e05
Change-Id: I2bbc08474daea7a019aca457e0725a4e9c2037bd