[roll] Roll infra/recipes [tree_closer] Ignore builders with magic property
Any builder that sets the `_tree_closer_ignore` property to True will
now be excluded from consideration for closing the tree, even if it
fails many times in a row. The motivation for this is that some builders
are nice to have in the global CI console for visibility, but they may
fail due to breakages in codebases other than the fuchsia platform.

There's no way to set builder-level tags, hence why we have to configure
this as a per-build property.

I named this property `_tree_closer_ignore` with a leading underscore to
emphasize that it's not a "normal" property consumed by a recipe
RunSteps function.

Original-Bug: 60655
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/431966
Original-Revision: 99602b74b82a6766fe566a1ed226bc8a712c2162
GitOrigin-RevId: 8b7c9d20983689e24be4a299d9e0d23d1da7f322
Change-Id: Ia2dd3c15eb0377352d17968eaa690714b2de97c7