[roll] Roll fuchsia [sync-from-stem][jiri] Error on untracked branches needing rebase
jiri emits a warning, but doesn't abort if there's branches in
dependencies that have no upstream but aren't rebased. In this case,
print out our own warning along with the command to rebase, as it's
otherwise easy to not notice that some subtree is out of date.

If the branch has an upstream, or is already based upon the desired
revision, no error is output and the sync will proceed normally.

Unfortunately this has to be implemented by regex grubbing through
jiri's output.

Test: python3 -m doctest -v sync-from-stem.py
Original-Fixed: 60889
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/432275
Original-Revision: 8699a880fa48dd5cd60c64440327f25f65c9f964
GitOrigin-RevId: 10a1e36785b4973228dde9009e1088815e334ea7
Change-Id: Id844e521d9ad4c65a3ec228c573b9284e0676e85