[roll] Roll fuchsia [utc][time][cm] Enable UTC clock propagation
TL;DR: If there are any issues with UTC time, revert this change.

This change enables the propagation of a UTC kernel clock owned by
component_manager and kept in sync by timekeeper. This clock gets
propagated through ELF Runner to all components. Appmgr, devhost,
fshost, and sshd are all configured to propagate the UTC clock if
it is set.

If there are any issues with UTC time being out of sync, this is likely
the culprit, as processes spawned by a process that is unaware of
userspace UTC kernel clocks will not receive the same UTC clock as the
rest of the system.

This is expected to be a transient issue only, as UTC time within the
kernel will be removed shortly.

Original-Bug: 60519
Multiply: timezone-test
Multiply: system_recovery_tests
Multiply: usb-system-update-checker-integration-tests
Multiply: fuchsia-pkg://fuchsia.com/pkg-resolver-integration-tests#meta/resolve-succeeds.cmx

Test: Manually boot product and verify that the time is current.
Test: fx iquery show timekeeper.cmx has a small delta between utc_nanos
      and utc_kernel_clock_value_nanos.
Test: fx test component-v1-time-unittests
Test: fx test component-v2-time-unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430674
Original-Revision: a02911052005809febd13ef41176f5b38195b0c1
GitOrigin-RevId: 422ec9ab78e5fc53764dec4e9abc90a607cfc0e6
Change-Id: Ibe30706860d4ddf4827863570d9d67933843054e