[roll] Roll fuchsia [minfs] Relax the fsck check for purged inodes.
When Minfs grows, the new inodes are not zeroed, so the magic for inodes
will not necessarily be zero, so the existing fsck check for this will
fail.

This has started to appear now because FVM's slice size was recently
changed to be 32 KiB rather than the 1 MiB that it was previously and
when Minfs formats the volume it only starts with 1 slice.

Also:

  + Changed the slice size for the file system tests to 32 KiB.

  + Ram-disks for the file system tests are now filled with a non-zero
    value to prevent us from inadvertently depending on it being zeroed.
    This required changes to the ram-disk code to allow setting a block
    size when passing in a VMO.

  + Added a log message to FVM binding so we can see the slice size.

  + Changed the file-system tests to use stderr rather than syslog
    because it's easier to debug.

Original-Bug: b/169352760
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/431954
Original-Revision: 5c7509e9d2f35d9064b1d5c74ee287bdf56bfc5d
GitOrigin-RevId: a2a8a58484d295ce38f642b71db6a2b7ab8d8b42
Change-Id: I913f751cb571eed97e65a874354711adf9f42878