[roll] Roll fuchsia [storage][extractor] Add extractor library 1/n
The library in its current state provides a way to write an image
file. Storage software can use this library to dump the data that is
relevant to them.

The library is in rust. It provides C bindings(in c-bindings directory)
that can be used in C++ code base.

This patch adds only extent part of the library.

Test: Ran host side unit tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427715
Original-Revision: ff937d6c8e9c367d1abf1111bbc9c7213d27d4ee
GitOrigin-RevId: 30b22a040ca06399eac99181a631770fc35feda0
Change-Id: I36a67024ca57ee7dac29c76e868b6e14092990c8