[roll] Roll fuchsia [codec] Add const VMO access API to InternalBuffer
This API will be used by the AML encoder to grab a handle to the sysmem
allocated VMO for passing to the canvas driver, and could be used by the
decoder for similar purposes if needed.

For now test where other InternalBuffers tests are.

Test: amlogic_decoder_integration_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/432556
Original-Revision: dcf7ab2ab50f333bf72c2068d4d49f8d709e5922
GitOrigin-RevId: 00eab900279a977e6c9673e05ceaeb461707075a
Change-Id: If71eb2e28cae64b0087da70453528086b3e0db26