[roll] Roll fuchsia [tools][fuzzing] Gate undercoat E2E test on environment variable
This CL switches the gate for the undercoat end-to-end tests to be based
on the presence of the UNDERCOAT_E2E_TESTS environment variable instead
of the absence of the `-short` flag to `go test`, so as to unblock work
on `go test` workflows.

This is an interim solution only: ideally, as tracked in
fxbug.dev/47225, this test would run cleanly in all environments
including CI. However, since it involves launching qemu against a
fuzzer-variant build, it is currently only usable manually.

This CL adds a comment explaining how to run the test manually, and also
fixes a few places that have bit-rotted due to the lack of CI coverage.

Original-Fixed: 58836
Test: cd tools; go test ./fuzz
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/422113
Original-Revision: 748743c6b84d2196c78d5acc1dd21c7414c9f6df
GitOrigin-RevId: 0f38b4e4e2fa964b3dd3d67539e5bc136329cc19
Change-Id: If80b8eb555ee6e5ad5678bf4bd11e03a68556258