[roll] Roll fuchsia [uuid] Make //src/lib/uuid host-compatible
The original //src/lib/uuid called zx_cprng_draw to generate random
bytes for UUID, which is not availabe on the host. In this change, we
use std::random_device to generate the random bytes on the host.

Note that in this CL we also modified
//zircon/system/ulib/gpt/BUILD.gn
because it has to modified together with this change. Otherwise, either
`fx build` or `fx set` will fail.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/432452
Original-Revision: e0bcb459968b08f6c07cc7b6d031cc99ef770fe4
GitOrigin-RevId: f94f59e196329fc65afeb49cca4e72f832778f42
Change-Id: Id3cdd8760f2ea72b43c696ff5f694c80cabdab04