[roll] Roll infra/recipes [3p_prebuilts] Use . for empty dest name.
The recipe uses a destination name read from a manifest as a step name.
However, one of the destinations is an empty string. Running this recipe
with bbagent causes the build to fail for the empty step name, so I
changed it to "." instead since the empty destination just means to use
the staging_dir directly.

Original-Bug: 60621
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/432198
Original-Revision: 8d50ba23c0984bc6b06603e874e4f2b532c5f506
GitOrigin-RevId: 915ed76ab33857bee3147e0ff09cde97700a792a
Change-Id: I415ffc12a6138ea593c629160f5fa988d6f6a43b