[roll] Roll fuchsia [swd-e2e-tests] Add packages.json support to util
As part of fxb/50750, support for the legacy "packages" file format in
updates is being removed in favor of "packages.json".

This change adds support to the util package to parse the packages.json
format[1] and fuchsia-pkg URLs[2] through an API that returns data in
the same format as the existing "ParsePackageList" function. A future
commit will remove this latter API. This change also introduces tests
for both APIs, and in particular that the APIs are compatible in the
returned data.

Original-Bug: 50750

[1] https://fuchsia.dev/fuchsia-src/concepts/software_delivery/update_pkg
[2] https://fuchsia.dev/fuchsia-src/concepts/packages/package_url

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430711
Original-Revision: 26efb3299465d66be68890812d8e879a288b676e
GitOrigin-RevId: 7e28e59e61719615d1f767616bf91dd695d21f6a
Change-Id: I88ecf1ed0e2b3bba294e89caed76a039c1e43700