[roll] Roll fuchsia [syslog] Don't stop streaming until context canceled
I'm not quite clear on when log_listener is supposed to exit, but we
definitely don't want to stop streaming syslogs if this happens. And it
does seem to happen sometimes - I wrote fxrev.dev/428815 to have
botanist fail if syslogging stops because of a log_listener successful
exit, and one of the tryjobs did fail because of that.

I also rewrote the syslog tests to eliminate shared state between tests,
which makes them easier to reason about and makes hard-to-debug failures
less common.

Finally, I renamed the syslog tests target from just "tests" to
"syslog_test" so that it can be referenced as "syslog_tests" in test
multipliers.

This should resolve the syslog truncation described in fxbug.dev/55301,
which happened because log_listener preemptively exited with a zero
retcode, causing the syslogger to quietly shut down. Now, we'll rerun
log_listener when that happens.

Multiply: syslog_tests: 1000
Original-Fixed: 55301
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430705
Original-Revision: f4a01f4de63cc909c7383e6e4224760b1971a7b4
GitOrigin-RevId: c6602040655cca64438b53da619cb2301796aeb9
Change-Id: Iada5f2c6e41ca3c463c4f046c20aa3590ca6bab2