[roll] Roll fuchsia [build] Clarify the contract of binaries.json
The documentation wasn't clear about the fact that prebuilt binaries,
which are not produced by the build, should also be listed there
(and that in this case, their paths cannot be used as Ninja targets).

For context, see [1] and the comment that explains the contract in
more details (explaining why the CL was later reverted).

Original-Bug: None
Testing: No change to the build.

[1] https://fuchsia-review.googlesource.com/c/fuchsia/+/431295

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/431298
Original-Revision: 759f246ed63ed1ecf82ea804a5c728a3e219fcf0
GitOrigin-RevId: 2e14c2c12ab4379460992af9fcde9ea6d11e509b
Change-Id: I1488b132a0644a8a44dd28dd9365073e9f37147f