[roll] Roll fuchsia [audio_core] Use sinc sampler if we might micro-SRC
To reconcile a client clock with a device clock, we might adjust
either of these (if they are adjustable); otherwise we insert a
micro-SRC frame-rate-conversion factor. Even if the device clock
supports fine-tuning, it may already be controlled by another
so we might use micro-SRC in that case as well. Only when the
client clock is adjustable ("flexible") are we guaranteed to NOT
need micro-SRC for synchronization.

Going forward, EXCEPT for the flexible clock scenario, we always
use the WindowedSinc resampler, as we may need its higher
fidelity when applying micro-SRC for clock synchronization.

Test: pending
Original-Fixed: 60061

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427561
Original-Revision: ac67f4f034e99dd6f84f64e9608102b960c6372c
GitOrigin-RevId: 6dc47184c89d0bec6c7e5b4b34de1dc71f0b4eee
Change-Id: I6ee0846c2d611ac4aa413e0081b4ad63b0ee8c42