[roll] Roll fuchsia [docs][clocks] Flag clock_get/clock_adjust as deprecated in the docs.
Flag the zx_clock_get and zx_clock_adjust as deprecated in our public
documentation as part of starting the process of removing these calls
from the kernel ABI all together.  Also, in clock_get, document the
"new" and "proper" way for a user to gain access to this information
in the new world.

Also, as part of clock-doc cleanup, provide a section in the
discussion of kernel clock objects which makes it a bit more clear
that the error bound reported by a clock does not come with any
guaranteed confidence interval.  Users who care need to know where
their clock came from, and what their clock maintainer's policies are
on the confidence levels of the error bound estimate.

With this in place, we can start the final burn down of removing all
of the calls to zx_clock_get.  If users have questions of what they
need to do instead, they can refer to the public docs for options.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/432503
Original-Revision: 138be977c33692f02a393437895a44d361efae01
GitOrigin-RevId: 9e3d1887e87b7725e1e9e16d2c66cfcd53191729
Change-Id: Ic15804a267e2e7cdf553286b07e3d3cdbb31c68f