[roll] Roll fuchsia [audio_core] Improved micro-SRC
This CL eliminates oscillation in the micro-SRC by quantizing the
adjustment rate so that it is a PPM measurement, just as the inputs
to the clock reconciliation process are.

In doing so, this CL also brings forward some of the reorganization
done in the coming "flexible" clock CL (for example, in PidControl).
As a result, the flexible CL will be just additions for that mode.

Clock tests using MixStage are split into their own unittest file.

Test: existing unittests reworked and tightened up

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/431080
Original-Revision: 8116b052b58119f41f477dd4a770dca9f433adc4
GitOrigin-RevId: e55a2686bcf66ed0cdab7bb5ee49a593221c5cea
Change-Id: I597c78d411f4281159dfdf1f10cf4c2f68a75956