[roll] Roll fuchsia [rust] Use a slice for channel write() and call() handles argument
Currently, the signature for channel write is:
pub fn write(&self, bytes: &[u8], handles: &mut Vec<Handle>)
  -> Result<(), Status>
However, slices are more idiomatic in rust for input arguments
because of deref coercion, which allows both Vec and non-Vec
types as input.

The reason Vec is used is to invalidate the Handle list
(with .set_len(0)) when handles are written to the channel,
however it is also possible to do this by directly setting
the items in the list to Handle::invalid(), which does not
require a Vec type.

However, this:
- changes the API to not resize the input
- has slightly more overhead in performance, though this seems
to be negligible and perhaps not worth optimizing if it isn't
a problem

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/429815
Original-Revision: c85f94128c419c1aabf733def18e15ccc596d66b
GitOrigin-RevId: 34939d6423699bc295ea11df3b9e6d32d6977e84
Change-Id: I4f55696a3479a4e4b05bb066bfd291badc38dcb1