[roll] Roll fuchsia [usb-hub] Fix memory leak in CancelAll
Our new queue-based system for tracking USB requests
didn't account for control transfers that were pending
in CancelAll. Added an atomic counter to track these
and wait for ownership to be transferred when CancelAll
is invoked. Eventually we will make CancelAll asynchronous
which should be the long-term solution.
to this problem.

Original-Bug: 60197

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/431834
Original-Revision: 2e18f162dd469861d692eeb7258958959f447220
GitOrigin-RevId: d1297cb99cf44fe2e9c4e0582f7606e8c5294a7e
Change-Id: Iac0f045f237a43a5bc858e18d9521ab05a692499