[roll] Roll fuchsia [build][alln] Migrate //zircon/system/ulib/zircon:zircon-headers to Fuchsia build.
This CL really allows building the Zircon VDSO (i.e. libzircon.so)
directly from the Fuchsia build. However, this is currently only used
to handle the .../zircon:zircon-headers target.

Migrating the .../zircon:zircon target entirely requires fixing the
sysroot setup to not depend on the Zircon legacy "zircon" target and
will require more work, for a future CL.

It is possible to build libzircon.so directly with:

  fx build user.vdso-arm64-shared/libzircon.so

The (stripped) result is byte-identical to the Zircon-produced library,
except for 8 build-id bytes. This was verified by comparing the binary
values of the two files, using the following command:

  diff -burN \
    <(xxd out/default.zircon/user.vdso-arm64-clang.shlib/obj/system/ulib/zircon/libzircon.so) \
    <(xxd out/default/user.vdso_arm64-shared/libzircon.so)

Original-Bug: 58364
Testing: Should not change the build.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/432696
Original-Revision: 2c817e306e6d50a7ef7c691e07cf66ecb5eacb70
GitOrigin-RevId: 80bdaa8811d3f031b202e7412ba85fe5705c0510
Change-Id: Ic723a06fd8b40a09370150e2f78b98fc7dc61025