[roll] Roll fuchsia [fdio] Enable retry of send/recv on connecting stream socket.
When the socket is still in connecting state, enable the callers of
stream socket read/write to retry the request after the state becomes
connected. This is as per RFC793, page 56, 58, for TCP send/recv.
In fdio, we have information about the socket state (connecting or
connected) and that can be used to handle this requirement.

Original-Bug: 39041
Test: fx test -od netstack-bsdsocket-c-api-test -- \
	  --gtest_filter=NetStreamTest/BeforeConnectTest.BeforeConnect/*
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/429554
Original-Revision: 1e4a85d327bc55f5ad383eac37bb139d1f43358f
GitOrigin-RevId: e5668c7eef4ac20a16daca1988a5f7b46c017702
Change-Id: Ia5a452bf3a992414eed9a039545d92e2a12750fc