[roll] Roll fuchsia [dart] Add dart tools to tool_paths.json
This will expose the paths of dart tools like blobstats to consumers of
the build, such as recipes. This is necessary in order for the recipes
to invoke blobstats using executable from the output directory, rather
than invoking blobstats' main.dart as a script using the prebuilt dart
interpreter. This will mean that recipes no longer need to know the path
to the prebuilt dart interpreter or to blobstats' main.dart source file.

Original-Bug: 58254
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/433194
Original-Revision: bcb43cde9d414da4c1682aee78054f1b82124257
GitOrigin-RevId: 9fcf1a731fb774478d3a8ba5422269ff189abf7d
Change-Id: Ief7f5a93f7ea3a278038df67c66f83bf9de37608