[roll] Roll fuchsia [tools] Add blobstats to //tools:tools target
We'd like to have recipes start consuming the blobstats executable
produced by the build, rather than executing blobstats as a script, so
we need to hook it into the build graph.

Once this transition is complete, we can move the blobstats source code
from //scripts (which is supposed to contain code that's only run by fx)
into //tools, which is a more standard location for code that's also run
by infra.

Also, specify $host_toolchain for other dart tools, since they are only
intended to be run on the host.

Original-Bug: 58254
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/433326
Original-Revision: 43999ef0913beff02fb7a2266dbbcc1bf1b5cf67
GitOrigin-RevId: de6d49b589d79eb0045eb4b48af499a412c9fc48
Change-Id: Ic90263c1268f1735a2535de3d3360b57560cec41