[roll] Roll fuchsia [fidl][gidl] Correctly format float32 literals
To avoid implicit conversion warnings when emitting float32 literals
make sure they're formatted as such with the "f" suffix.

This change adds "f" to the end of float32 literals if there are decimal
points in the string representation of the literal.

Original-Bug: 58162
Test: built with tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/432569
Original-Revision: 0227a1fcaae037b3060288da72d4fb28464548c5
GitOrigin-RevId: 892538e45f5b02160d5fb74b02584022c89b5eb5
Change-Id: I06c1454c6c4b8de2c7ccb81fe92825d93fe73408