[roll] Roll fuchsia [bootserver] Improve file transfer logs
I noticed that after the bootsever finishes sending a file, it currently
prints a "botanist done" log. This is confusing because it implies that
botanist itself has completed its work, when in fact it's only finished
transferring one file.

We can fix this by using Debugf instead of Infof for this log. Infof()
doesn't print the log level, which is why it's just "botanist done" –
with Debugf, it'll instead print "botanist DEBUG: done". I also changed
the "done" log to be more verbose and actually say which file just
finished being transferred.

Finally, I deleted newlines from the end of the logs, because they
aren't necessary as long as we're using the logger library.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/433434
Original-Revision: d71be0958679fb596dbaea30353b0fc00a502361
GitOrigin-RevId: 0f8a48b5623310e0ef94fe9786503447a3def499
Change-Id: I9e3c68f5c02d9f40dad555afeaab64c5226c71ba