[roll] Roll fuchsia [microbenchmarks] Reduce time-limit-exceeded flakes by reducing iteration counts
microbenchmarks_test has been getting flaky failures in which it
exceeds the time limit for tests (the default for which is 5 minutes).
This has been happening occasionally for a while, but recently the
rate of occurrence increased a lot.

The reason for this is rather odd: microbenchmarks_test is taking a
lot longer to run on some variants of the terminal-x64-release builder
than others.  On one it takes 1.6 minutes (though with longer times
recently), and on another it takes 4.5 to 6 minutes.

A spot check of the performance test results shows that two test cases
in particular are a lot slower on one builder, though I don't yet know
why:

 * Fdio/Spawn
 * RoundTrip_Futex_SingleProcess

For now, reduce the flakiness by reducing the overall time taken by
microbenchmarks_test by reducing the number of iterations by 20%.

Original-Bug: 59553
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/402426
Original-Revision: c5f4898991a247788d795db06d8b2682a50d5899
GitOrigin-RevId: 81719ffb564a883d7cc95f5f377d9ae782dac127
Change-Id: Iad64f582bb6a32432d68bff10ded2889c911a751