[roll] Roll fuchsia [storage][extractor] Add extractor library 2/n
The library in its current state provides a way to write an image
file. Storage software can use this library to dump the data that is
relevant to them.

The library is in rust. It provides C bindings(in c-bindings directory)
that can be used in C++ code base.

This patch adds only extent cluster part of the library.

Test: Ran host side unit tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/432634
Original-Revision: 5d926acf7a406a97a7609d9005f1b45aedcdf1cd
GitOrigin-RevId: b88073aaf9561c946bf3732068efb320b2a77c10
Change-Id: Ibbac86d31805c000c7212c6f266c788e47ff3205