[roll] Roll fuchsia [fx][formatting] Select and build the right doc-checker when formatting
This fixes a bug in which the `fx format-code` was not consulting
metadata to figure out which doc-checker to use (it hard-coded a path
under out/default). We update the formatting script to rely on
tool_paths metadata to figure out where to find the tool.

We also superficially update the fidl-format logic as it is inheritently
similarly and so it would be nicer to have a consistency among the
names/spellings/etc. across the two cases.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/433974
Original-Revision: 7b1def2207147d2da8e938e7f71b61a5eef680da
GitOrigin-RevId: 39d8382e983cafe6bd9cd93c8ac6c42e32c1351a
Change-Id: Ic244290800714ca750094ef5999adf5aa392f36f