[roll] Roll fuchsia [timekeeper] Add a clock manager and estimator.
The estimator is a placeholder within which the kalman filter and
frequency tracker can live in the future (without requiring any async
code!!). The clock manager consolidates some more of the logic
previously in main and will be the site of the clock update logic once
we start to use slewing. As of this CL there should be no change in
behavior of the clock or of the diagnostics.

Original-Bug: 56858

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/432625
Original-Revision: b597d62941ffa442de709c5126b956d8e6460c20
GitOrigin-RevId: e7bf6f46f11a222f5281acfb4a5d39efe5ad4e46
Change-Id: Ia85448edc06992314c5100036f51024980e40c8e