[roll] Roll fuchsia [devmgr_launcher] Remove access to /boot and /system
All tests which depend on this library should be packaged and avoid
access to /boot and /system in order to be hermetic. As a result we can
assume access to /pkg and simplify our library interface by not
requiring packaged test to specify path_prefix.

This change does break the ability to write an isolated devmgr based
test in a bringup build for now, however at some point in the future,
this capability may be restored if we add support for a component based
test runner in bootfs.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409220
Original-Revision: 7a3ea2b1a66cc6d46273e57c83a3e864812c2f34
GitOrigin-RevId: 555538e88d5dc2f116ca78527dc88b1b9fe67d4a
Change-Id: Ia8d2756abcb00682b0bf4c50c1efe2e18da6771c