[roll] Roll fuchsia [time][flake] Fix timekeeper integ test flake
To avoid faking out the entirety of the network state service, the
timekeeper integration test injects a fake that does only enough to
allow the setup of the network check to succeed, but not the entire
check. This is okay because timekeeper tries to sync time anyway even
if the network check fails, so long as the setup succeeds.
However, the setup involves making a request through the connection
made to the network service. If the connection is broken before that
request is made, timekeeper crashes and the test fails.
This change ensures that the channel held by the fake is held until
the required request is made.

Multiply: timekeeper-integration: 30

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/434044
Original-Revision: 9d73dcc34b796f844437401f84ed79046e54eaa5
GitOrigin-RevId: 28c5a30fd21aed0f62a8c9e93d718c27a25182a8
Change-Id: I8fe092b36054763804cb68ca9a7f1f2c054cb0be