[roll] Roll fuchsia [tools] Move blobstats to infratools
blobstats is used by the infrastructure, so it must be included in all
builders (//bundles:infratools is built by all builders). It always runs
on the same machine as the build so it only needs to be built for
$host_toolchain.

Also clean up generation of toolchain-specific tools, and stop
building x64 and arm64 versions for some tools that only need to run on
the builder: artifactory, testsharder, and size_checker.

Original-Bug: 58254
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/433814
Original-Revision: 2ebb9cee6c57d9a8243905861f2a0ac69ddd6594
GitOrigin-RevId: 76cee16cb681ab4f042d1d37d79ce9be52cd7bb0
Change-Id: I4b38e0820c11e3de3f875a284aeb9e08326020d8