[roll] Roll fuchsia [fvm] Replace some uses of FormatInfo.
Removes some users of fvm::FormatInfo with fvm::Header instead.

The most critical change is in block/drivers/fvm/fvm.cc. The resizing
code is simplified to compare slice counts for before/after (previously
it compared sizes of the allocation table) and the header updating code
is moved to a new helper in fvm::Header that updates both the
fvm_partition_size and pslice_count.

This new setter is also used in the fvm::Header constructor so there's
one consistent place where fvm_partition_size is set.

The VPartition code previously had access to the FormatInfo in the
VPartitionManager to do some computations. This FormatInfo member was
misleading because it isn't kept up-to-date, but does allow certain
computations to be done on this copy without locking.

This is now changed to a new GetHeader() accessor that reads the actual
current FVM header and returns it by copy to the VPartition. The
VPartition code that did this seemed to already do some locking so I
don't think this overhead will be significant. This also avoids
out-of-sync problems.

The slice size can never change and is accessed in several places so a
duplicate copy of this is kept that can be read without locking.

The FVM header constructors are fixed to take into account the metadata
size when saving the used slice counts. This fixes the
PreallocatedMetadataGrowsAsMuchAsPossible test.

Update the sparse image test to supply a valid slice size. Previously
this was smaller than the block size which is invalid and was caught by
a new assertion.

The block device test and FVM block driver are updated to take a full
metadata VMO rather than just a FormatInfo. Since it doesn't use
FormatInfo any more this is required, and should be less brittle since
the FVM driver has the full context available.

FVM resizing:

  The test FvmResizeTest.NonPreallocatedMetadataIsUnaffected is deleted.
  This test was wrong. FVM does not guarantee that a non-preallocated
  FVM can never grow, it just doesn't guarantee growth. The allocation
  table may have padding which allows for some growth. Previously this
  test passed because in VPartitionManager::Load() it did this
  comparison for resize which is comparing the allocation table size for
  the current disk size with the current allocation table, which is
  always the same:

    AllocTableLengthForDiskSize(GetFvmLocked()->fvm_partition_size, GetFvmLocked()->slice_size) <
              GetFvmLocked()->allocation_table_size

  FvmResizeTest.PreallocatedMetadataGrowsAsMuchAsPossible is fixed.
  Previously it would check the slices against the original slice count
  rather than the maximum possible one. Previously the test passed
  because of the above bug where FVM didn't resize.

  Rename kBlockSize constant in the resize test which was shadowing the
  global fvm::kBlockSize constant.

Original-Bug: 60920

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/432621
Original-Revision: 495e707c142027bb5bce88c99ea6853dcaec6a23
GitOrigin-RevId: 0f762b1615d626dd8c27478de4252f54977c28f0
Change-Id: Ib7f6de0c905681785cbd56fc3c63a468db1e1d13