[roll] Roll fuchsia [testsharder] Handle the case where test count < max-shards-per-env.
The broken builder in the attached bug is failing to create shards
because it's missing test durations so it defaults to the
maxShardsPerEnvironment for the number of shards to create. However, the
number of tests it runs is less than that, so it gets an indexing error
when trying to access the tests in a shard because some of the shards
ends up with no tests.

Original-Bug: 61212
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/434313
Original-Revision: a16708141f1a3d26018f9228181b9de3fd8e599f
GitOrigin-RevId: 2a1d135f7df8b7730f55b315b8b83ef5ccea1af5
Change-Id: I705b035839cca8fc5aad52859e32145ba275e398