[roll] Roll fuchsia [debugger] Use maximum module extent.
Compute the extent of ELF binaries and don't match symbols outside of
these bounds. Without this, random heap addresses can match the last
symbol in the nearest mapped shared library.

Add a constant limit on the distance away from ELF symbols we can match
addresses as a backstop for cases where this fails. The goal is to find
a symbol if it's there but not to be too confusing.

Updates the module symbol test initialization to assert that
initialization works. If some files are missing the previous code would
crash rather than report an error.

Original-Bug: 42243
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/434474
Original-Revision: d877f0008a24a518a49b364f486d19238a696492
GitOrigin-RevId: 7ea0427546ed39d62d4fa22024815d748ee4f37c
Change-Id: I9aba4486dc29af07b874b7fdff374673d8e15941