[roll] Roll fuchsia [audio_core] Update clock terminology
AudioCore is updated to consistently use just the "adjustable"
and "non-adjustable" labels for AudioClocks (in addition to
their source: "client" and "device"). Other terms are removed,
except the term "flexible" which is still used sparingly in
the directly-client-facing renderer and capturer classes.

In other CL, "flexible" replaces "optimal" or other
terms in documentation, examples and public headers when
describing the clock supplied (and adjusted) by AudioCore.
Although internally we refer to such a clock as "adjustable"
(from AudioCore's point-of-view), that would be entirely
confusing to clients since it is a clock that they CANNOT adjust.

Test: No behavior change, naming changes only

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/434102
Original-Revision: cad75d6d0ee0b1ba84d039c8ab4ee941af5d1834
GitOrigin-RevId: cefdb5da9f1bdadf2e0ffca20b04028e7075c19c
Change-Id: I57297310df3652f13177ca3416e6ecbd5ea91c86