[roll] Roll infra/recipes [auto_roller] Support additional labels
Support setting additional labels when pushing roll CLs. This is
necessary for repositories that don't just use CQ but have other
presubmit validation that requires a label to be set to trigger a run.
Also setting "Commit-Queue" to 0 suppresses setting "Commit-Queue" on
push.

Support waiting on additional labels set by non-CQ presubmit systems.
Will wait on these before checking to see if CQ has completed (if there
is CQ).

Not all presubmit systems can directly submit changes. After checking
that presubmit has passed and if this is not a dry run, check if the
roll has been submitted and if not submit it. This is the only change to
existing expectation files.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/422055
Original-Revision: d52d4ce8060450f451b393ba3fa314b6b16a0591
GitOrigin-RevId: bd01e2323d42a001755e0e73324bfe061541d4cf
Change-Id: I17982494b9beb8ca8b4238b0bdea0f0c1cdfc756