[roll] Roll fuchsia [lsan] Fixes LSAN memory leak in display controller
Fixes the lifetime of a watcher closure, which under some circumstances
could cause a reference to the closure not to be deallocated.  The old
code employs a clever trick to ensure that a reference to a closure
expires when the closure is called.  However, it does not release memory
if the closure is never called, leading to a memory leak.

The new code ensures that the reference is removed when either this goes
out of scope, or the closure is called.  The approach is not thread
safe, and is assuming that the async loop is attached to the main
thread.

Original-Fixed: 61246
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/434517
Original-Revision: c57af663194f521e77aabda7d254fea0369017c2
GitOrigin-RevId: 197bb892680e997a12cb207aa0fedcf1588db24a
Change-Id: I129dd084d8c02483a865967e2bd1fab171274aac