[roll] Roll fuchsia [logs] optimize reconfigure to init with config
By default, creation of a global logger within a process will attempt
to create a socket connection for use with LogSink protocol (assuming
syslog is not statically included). When used with fx_log_reconfigure,
this connection is potentially thrown away in the case where a new one
is provided via the logger configuration. This optimization allows the
initialization path to use a config provided when creating the new
logger and avoid the unnecessary steps of creating the superfluous
connection.

Original-Bug: 54198
Test:
  fx test //sdk/lib/syslog
  fx test //src/diagnostics/archivist

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/428831
Original-Revision: 03196e5d139bac22a7400a84a78720fbeb1e58ea
GitOrigin-RevId: 5cca468ca2a01001bbaa0f27ee36d9f9f6b21574
Change-Id: I625f2273abac000a0a7e7a113a47aaeae9996dc3