[roll] Roll fuchsia [kernel][acpi_lite] add new lightweight library to parse necessary kernel ACPI tables
Add a simple library that replaces the simple part of finding the root
tables and decoding the simple ones that are used in the kernel.

Next change will wire it up to the acpi tables library which has test
code validating that the low level interface is working.

Testing: validate that it finds the ACPI table on a variety of hardware
including qemu and a few AMD and Intel machines.

Original-Bug: 32149
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427694
Original-Revision: 645dbffdfbf52f72491187a1c59af1a677e3dd00
GitOrigin-RevId: eb201542d5cd503cd0bd064ab92362659e36ad32
Change-Id: Ied985226f31ae12b26fdf164db0f4939e034c492