[roll] Roll fuchsia [kernel][acpi_lite] Add C++ interface and add tests.
We move to a C++ style interface for acpi_lite, eliminating global state
and adding the ability to fake out the implementation of "phys_to_virt".

In turn, this allows us to add tests for acpi_lite. We include data on
3 platforms at the moment:

  QEMU v5.0.0 (Emulator)
  Fuchsia's Hypervisor (Hypervisor)
  Intel NUC 7i5DN (Hardware)

We additionally update the code to be more explicit about the size of
memory it is attempting to access: this allows us to validate that we
won't read past the end of memory, and in future allow us to implement
ACPI on platforms where we need to map pages in to access ACPI tables.

In this CL we retain the C-style API for acpi_lite. This is removed in
the next CL.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/428216
Original-Revision: e70677d5a639e4229ebc6827ec2157e06525ec9c
GitOrigin-RevId: 57b3c5142f0e3826728293162dadf99c9d1b0867
Change-Id: Ib986a2f94b6e08796d874928d74034fba34e5c49