[roll] Roll fuchsia [i915] Unconditionally verify AllocChecker result.
The current i915 code only checks the result of an AllocChecker in debug
mode. That is, we have the code:

  assert(ac.check())

which is optimised away in release builds. This has previously worked,
because the AllocChecker::check function was:

 1. Compiled out of line, in a different compilation unit; and

 2. Was compiled with different compiler flags than i915, which happened
    to mean that `check` was a no-op iff `assert` was compiled away.

The CL If2277bbaa5288175037667f50552d92c2509a2a5 changes AllocChecker to
generate more efficient code by performing most of the checks inline.
This means that the above two conditions no longer hold: in particular,
because the AllocChecker code is inlined, it now adopts the i915 flags
instead of the fbl library flags. As a result,
in some builds, the `check` call is compiled away when the AllocChecker
library still expects it to be called, leading to a crash.

The CL updates the code to unconditionally call `check`. This has
negligible overhead, and no overhead when ZX_ASSERT_LEVEL <= 1.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/435062
Original-Revision: 5189f15214ab01378021a2064601291309615708
GitOrigin-RevId: 20fda6e3f0a69fd0e380631629f4492150a91f08
Change-Id: Ia093da6a2eb221d23084b32438ada066becff875