[roll] Roll fuchsia [devices] Request private clones of driver binary
This is part of the drivers-as-components work. As of fxrev.dev/429697,
memfs will only set ZX_RIGHT_SET_PROPERTY on private clones of VMOs.
When requesting the VMO of a driver binary, request a private clone so
that we can set the name of the VMO for accounting purposes. As bootfs
is backed by memfs, without requesting a private clone we can not name a
VMO.

Original-Bug: 33183
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/435066
Original-Revision: 18f0420e42570b4645e1673fa0bffdd3c6190d46
GitOrigin-RevId: 4da13b30fbdec95bfc999fca9519158ebdc7ab37
Change-Id: Ib98d595654ff9ce46f9990c5a4a7731245985c2a