[roll] Roll fuchsia [build] Prepare fidl_c_header.gni for sysroot migration.
This CL fixes an issue in fidl_c_header.gni, which is used to
create the FIDL C header for a few FIDL libraries that are
still used by the Zircon build. This is required to properly
move the sysroot generation to the Fuchsia build in future CLs.

The root issue is that zx_host_tool() in the Fuchsia build does
not support, as its ZN-build version, a depfile that is generated
by a dependency. This results in Ninja wanting to rebuild the
fuchsia-{debugdata,io,mem}.fidl.d files and their dependents on
each invokation (when generating the sysroot in the Fuchsia build).

This fixes the issue by generating the FIDL C header and the
depfile in one go, using a custom script instead.

Inspection of changes to out/default/ shows that this changes
the build by changing the commands to build the header, and removing
now-uneeded stamp files for the depfile-generating target. No suprises
here. Finally, the content of the generated fidl.h files were compared
before/after the build to verify they are the same:

  gen/fuchsia-debugdata.fidl/include/fuchsia/debugdata/c/fidl.h
  gen/fuchsia-mem.fidl/include/fuchsia/mem/c/fidl.h
  gen/fuchsia-io.fidl/include/fuchsia/io/c/fidl.h
  gen/zx.fidl/include/zx/c/fidl.h

Original-Bug: 60613
Testing: See above.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/434755
Original-Revision: c7cc83a0f664c138f3ba06ba4908c04ce3b3d4a4
GitOrigin-RevId: 6fa0b7b2980cfa9e2a970351a54a48675afd1c63
Change-Id: Iaa2f93c370dbf8ea528f244b514de7fcce9fade0