[roll] Roll fuchsia [sys/core] put storage in "persistent" subdir
Modify the storage capability declaration for "data" to use a subdir
with the name "persistent".

This is a breaking change, and users of this storage capability will
lose access to their old storage due to this change. Archivist is the
only such component, and can handle this.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/434194
Original-Revision: b5ea2ccd4fc7da406925b675e4e212d519608e25
GitOrigin-RevId: 54a34ff52629d4aed4588890b9c32d342f1a02d9
Change-Id: I5b97d9faa165432e5941a550c60667bb17700a0f