[roll] Roll fuchsia [wlan][sme] Log previous connected AP to Inspect
If a bug report is uploaded while a user is not connected wirelessly, or
while a user is connected to a different AP, we would lose information
for debugging the previous disconnect. In this commit, we log to Inspect
the previously connected AP once SME moves to the idle state from a
connected state.

Original-Bug: 60154
Test:
1. Manual verification
2. fx test wlanstack-tests
3. fx test fuchsia-pkg://fuchsia.com/wlan-hw-sim-test#meta/verify-wlan-inspect.cmx

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430314
Original-Revision: a18343eaf1617da64ecfaddb297f3c01f66f937c
GitOrigin-RevId: 76c2762f456c5699e6396c0032d9f0ff06dbcd3c
Change-Id: I028c0c232ebb19379b1aae1244262b8ad948fc84