[roll] Roll fuchsia [flatland] Default FlatlandPresenter impl.
The FlatlandPresenter interface was primarily added to:
1. Isolate Flatland from having a direct dependency on the
   FrameScheduler.
2. Introduce a natural location to add a thread-safety later around
   the FrameScheduler (rather than enforcing thread-safety within the
   FrameScheduler itself).

As a result, the simplest implementation of the DefaultFrameScheduler
simply places FrameScheduler calls into a task posted onto the main
looper. A lock would be simpler, but GFX clients do not access the
FrameScheduler through the same object, meaning a lock is not
sufficient when Flatland and GFX are running simultaneously.

Future changes in the FrameScheduler could eliminate the need for the
additional thread-safety logic.

TEST=fx test -o flatland_unittests
MULTIPLY: flatland_unittests (fuchsia): 30

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/434038
Original-Revision: 18386de15348582c2ad05604a381a9a564461bd5
GitOrigin-RevId: 744182c7994046bca7d35e163f29fe7b9c07e77a
Change-Id: Id382919c3713a29f6ccb79303c5832301446128a