[roll] Roll fuchsia [ffx] Drop futures when FIDL client stream drops.
This fixes up a bit of a hack from a previous CL so that there aren't
any wrapped errors. This is identical to `try_for_each_concurrent`
except that futures are dropped when the client disconnects and the end
of the stream is reached.

This does so with `try_for_each_concurrent_while_connected` which
handles requests so long as the end of the client stream has not been
reached.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/435055
Original-Revision: f7bd8b5ed4d4ce93fde1fe24f297cb610fb40fb2
GitOrigin-RevId: ece5d38ecd432521a875c81bd2652095fc9ce41f
Change-Id: I596175cf8cbb675edcd4b8488f4f02e155cd50cb