[roll] Roll infra/recipes Remove all `os` imports from recipe_modules
`recipes.py lint` disallows imports of `os`, but it currently only works
for recipes, not recipe modules, so we still had several `os` imports in
our recipe modules. I was able to substitute all direct usages of the
`os` module with usages of the corresponding `api.path` or `api.context`
attributes.

This also fixed a pylint warning in the build recipe module, which has
several methods that take an `os` argument. Pylint was giving a
"redefined-outer-name" warning because these arguments clobbered the
imported `os`, but that warning is fixed now that we no longer import
`os` at all.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/435535
Original-Revision: 645882223519a0aa0799b7bae2350afa677a8592
GitOrigin-RevId: a13fd3913ff7633358e69df7b8340c5cd8740632
Change-Id: Iddd41009ee4d2a025da7b511def4f950f5314be5