[roll] Roll fuchsia [core][bringup] Support appmgr resolution failure
Disable the package resolver in the bringup product since bringup
is not expected to support blobfs and pkgfs.

Introduce the 'core starter' component which is in charge of
starting appmgr. Without the package resolver, resolution of
appmgr fails. With the previous topology this was fatal to the
system because appmgr was in a direct `eager` lineage from the
root component and its failure to bind cascaded to the root.
The core starter becomes the direct, `eager` descendant of root,
attempts to use a capability from appmgr, and exits, regardless
of result.

Original-Fixed: 58182

TEST=startup-integration-test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/428040
Original-Revision: 124f955ae0d1db1c7e991684c7e8a9b4528d6806
GitOrigin-RevId: 1b1fdbbca7130009e11068ada4cc25f8ec684957
Change-Id: I770bd8a95be01baee5772f4557f35c626638eedb