[roll] Roll fuchsia [i915] Trivial: Simplify `Controller::AddDisplay`.
Currently, we reserve a single slot in the fbl::Vector each time
`AddDisplay` is called. Each such `Vector::reserve` operation will need
to perform an allocation and move all existing elements to this new
allocation. The result is that adding `n` displays currently requires
O(n**2) operations.

While the number of displays is likely to be small (so n**2 is also
likely to be small), it is also simpler to just not reserve the space.

Simplify the implementation a little.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/435063
Original-Revision: 08da000905ecfc6d8f99ec582ae5b0ca5098b670
GitOrigin-RevId: 8dbc0debea12d41c0734c107392f5f8d9f4d10f3
Change-Id: I0ff3d2c57ae8f124bd9fae6d3c9d00d10a6348fc