[roll] Roll infra/recipes [tricium] Handle deleted lines.
A deleted line gives a line_count of 0 which then causes the diff range
to not be recorded. However, if that deleted line causes formatting
issues, this will cause a failure because it tries to intersect the diff
range of the formatted file with the diff range of the changed file and
it can't find the diff range of the changed file because it wasn't
recorded in the matches.

Original-Bug: 61384
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/435534
Original-Revision: 1cf9fcf52ff3e0342a8da766e46459b84140466f
GitOrigin-RevId: 7395441297c31918cef248b772575bf9ae98fbc0
Change-Id: Ib216bb192a2988c0b7712eec9f6f18e7b351c6ca