[roll] Roll fuchsia [kernel][arm][gicv3] Read the correct max SPI field
Previously was reading a field that means something else, but was
probably always larger than necessary on all of the socs that have
been tested on previously.

Updated the GICv2 detection to print a similar message.

Tested: boot on a few pieces of hardware, validating that the interrupt
count that is read from the register lines up with reality.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/424033
Original-Revision: 8779230d67a47860be45cd82bf59b72b94f5db79
GitOrigin-RevId: 6d4739e45ecde69608d3ed5620b80b2e4aa75097
Change-Id: Id3bc78268696f21176358e42a21aaaa73364b443