[roll] Roll fuchsia [audio_core] Fix distortion caused by gaps between mixes
See fxbug.dev/61128#c9 for a detailed description of the bug. This CL
implements the same caching of partially-read data that we previously
implemented for EffectsStage.

I've had a very difficult time getting the bug to trigger in an
integration test, including the vendored integration tests. This CL
adds some additional integration tests, but none of them fix this bug.

I've manually verified the bug is fixed: Before this CL, the bug would
repro as described at fxbug.dev/61128 within 20s. After this CL, the bug
did not repro after 1000s. I would like to get this fix in while working
on an integration test in the background.

The CoefficientTableCache has been updated to precache 2->1 and 1->2
sampling ratios, which are needed to avoid underflows in the new tests
added to audio_renderer_pipeline_test.cc.

Original-Bug: 61128
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/435141
Original-Revision: c7faf90a9111173279692f26188284d50f392a73
GitOrigin-RevId: ab42eb49b162e34c358383f9276e71e4b72fc028
Change-Id: I3181de7890f3acdfa3ed51a9e41df7843a4c5bc0