[roll] Roll fuchsia [kernel][oom] Ignore dlog_shutdown result when OOMing
When rebooting because of an OOM condition, don't assert on the result
of dlog_shutdown.  Failing the assert results in a kernel panic and
misclassification of the reboot cause.  Two possible reasons for
dlog_shutdown to fail are 1) the dlog subsystem is already in the
process of being shutdown and 2) CPU starvation has prevented the
already queued dlog messages from being written to the serial port
before the 20 second timeout has elapsed.  In either of these cases
there's not much else we can do that's better than simply continuing
on with the reboot.

Fix and update adjacent code comments.

Original-Fixed: 55673
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/435435
Original-Revision: 4d4f919ba809478768616ec88d26278ca73eeb01
GitOrigin-RevId: b640efc158f76e31263f5d6c4c72acbaa73c529a
Change-Id: I2f66de393df8e0ce8815655f2b34797e0a9abd07