[roll] Roll fuchsia [fx] Make `fx list-build-artifacts` building explicit
Now that the name has been changed to "list", the implicitly behavior of
building is awkward. This change replaces `--no-build` with `--build`.

This will require an update of a local checkout w.r.t to the
work-from-home workflow, but that is already required of the parent
renaming change, so it minimizes disruption to do this now.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/435575
Original-Revision: 9e3b40a85d9996dd84f96018c106694bb4b3fcf7
GitOrigin-RevId: 3e7d9f5d2b8289a97a9461fe8da9fcb4e6225a24
Change-Id: I7074f1faa2c38f9401b5318d219d7fa34add4d72