[roll] Roll fuchsia [fx] Move flash-remote over to new fetching logic
This change moves `fx flash-remote` over to using the means of fetching
remote build artifacts that `fx pave` and friends now use. A side-effect
is that we are no longer reliant on the build archive - and are rsyncing
precisely the set of images that flashing requires.

Original-Bug: 58096
Test: updated fx_flash_fuchsia_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/423199
Original-Revision: b895334fa9d619f1899318f8f7b145e1460c487f
GitOrigin-RevId: c2e9cc264a2755b9250b54a94fb9028fc3c7e8dd
Change-Id: I1e3e6854339faa9ee03738208c8534084cf9cfc0