[roll] Roll fuchsia [wlan][policy] hidden network config support
This change adds a field for the estimated probability that a saved
network is hidden. For now the value will always be 0, meaning we
shouldn't actively probe for the network. Another field is also added
to network configs for deciding whether the network is hidden: have
we connected to this network using a passive scan. It also adds
methods to the SavedNetworksManager to update configs.

Multiply: wlancfg-tests
Original-Bug: 35916
Test: Added unit tests
	fx test wlancfg-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430795
Original-Revision: 0758d9264225eaa2d00da7f589d1f3f2a4052838
GitOrigin-RevId: fdd8ab2e705d6c8d4b49ba34fbfd9e576206df26
Change-Id: Ic661e416d5bb535a4b110f9ce7918dfefd92a031