[roll] Roll fuchsia [doc_checker]Allowing http link to OWNERS files
This file is in the //docs path, and the general rule is
http based links are not allowed to files under //docs, they should be
file paths starting at /docs.

The OWNERS file is a special case since it is not markdown, it is
not copied to fuchsia.dev for publishing, so it should be a link
back to the source code.

This fix adds a list of allowed paths to handle this situation.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/436054
Original-Revision: 218aaa3126ac09c6526205160908a314ee89cad4
GitOrigin-RevId: bdb0984cbef73a19bbd58c28d4ea5d406ed30913
Change-Id: I430bb862baf17c446913c6944f81fb3434d6404d