[roll] Roll fuchsia [build][fuzzing] Fix logging for no-Go fuzzers
fxr/412795 encountered trouble building go fuzzers with its changes to
the syslog backend. It tried to work around this by building the
libraries as static when in a fuzzer toolchain. This did build, but the
resulting fuzzer binaries had weak, unresolved symbols for the logging
macros, leading to crashes.

This CL isn't a fix for the Go fuzzer linking issue. Instead, it narrows
the workaround to only apply to go fuzzers, and to forbid them from
having a dep on //sdk/lib/syslog/cpp. This allows no-Go fuzzers to
depend on syslog in the expected manner.

Bugs: 44551, 59937, 60948, 61348
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/436074
Original-Revision: 157405c63252dc58c38b2359309e0afae0d3e4f9
GitOrigin-RevId: d2b4de8ca7b96cd4687656834038bd20b515ad23
Change-Id: Ie69b5ed01069aa82dcde728335332125426866d9