[roll] Roll fuchsia [tools] Use correct target for triage and add tool_paths metadata.
The previous target was always installing the tool with the host
toolchain, not the linux x64 toolchain that we want. Also, the
tool_paths metadata was only getting added if the current toolchain=
host_toolchain, so although it would build the linux x64 version of the
tool, it wouldn't record the path in the tool_paths.json. This change
will add the metadata if the current toolchain is for any host, not just
the current host.

Original-Bug: 57570
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/436195
Original-Revision: 0b6b39dce14da7fb59faa95d15490d4033090587
GitOrigin-RevId: f3704c6096a42e1731b8b0a46bf49250b1d44a33
Change-Id: I1966fc821e83800ed2587a32a1d56ad2610e3231