[roll] Roll fuchsia [gidl][hlcpp] Separate FIDL generation per GIDL file
Currently, the HLCPP benchmarks are built within the same
source set, with each benchmark importing a large file
with all of the FIDL headers. This follows what is done
for the rust benchmarks so a minimal number of FIDL files
is used for each GIDL file.

This does not have significant effects on build time, but
it does move towards a leaner build configuration and the
output filenames now match the specific GIDL test -- making
it easier to understand what is going on.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/435843
Original-Revision: 284f05c74c82e98cbee4ad8fc1d2791e59e1c9ba
GitOrigin-RevId: 798fe7e783c7eec53293e13992bba9d4159136b7
Change-Id: Ib782ca34b7aa329bb3b0f4ebf8d6efb10b2eb8ef