[roll] Roll fuchsia [blobfs] Implement the BlobLayout class
The BlobLayout class defines where the Merkle tree and data are
positioned within a blob for different formats.  Additional methods are
defined to consolidate the calculations performed on the layout to a
single testable location.  This avoids having to worry about underflows
and overflows when dealing with the layout in the rest of blobfs.

Original-Bug: 36663
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/429216
Original-Revision: f9888b2fa4db73107699f47f204aafa86d81b881
GitOrigin-RevId: 1dc815c5974d5d4fc904c5161ec70e117a3c3c73
Change-Id: I2b1b3dc111fbf566c074209f3bb7a3fc3e57bc7e