[roll] Roll infra/recipes [build] Don't skip building if recipe_testing.enabled is true.
The recipes builder sets that property to true for all the builds it
launches. In these cases, we want to test all the builder steps, so we
don't want to skip the build even if the CL it is run with is determined
to not affect the build.

Original-Bug: 61506
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/436474
Original-Revision: b14f34273e216dab9d7d42b32e181bb8a5085b24
GitOrigin-RevId: c9cb5a2a4ca83af7f9f2efe13dfa57b823ead244
Change-Id: Ifb98b4e5b907f388c5ffa596bda8a3e26ecda3ad