[roll] Roll infra/recipes [build] Add support for running `fint set` instead of `gn gen`
This introduces three new fuchsia.proto fields: static_spec, fint_set,
and fint_build. If `fint_set` is enabled, we'll build fint and run `fint
set`, passing it the specified path to the static spec, instead of
running `gn gen`. Eventually we'll run `fint build` instead of `ninja`
if `fint_build` is set, but I saw no need to implement that code path
yet since we're nowhere near having `fint build` implemented in the
platform.

We had originally planned to not have a `static_spec` fuchsia.proto
field and instead have the recipe dynamically calculate the path to the
static spec based on the builder name. However, that would require that
the recipe know the location of the static specs in the checkout, which
would make it more difficult to change the paths of static specs in
integration.git, or to modify the mapping from builder to static spec
path if and when we want to have builders share static specs. If we
instead calculate the `static_spec` in the infra configs, we can easily
change static spec paths in a single integration.git commit, with no
soft transition required.

It's not easy to test this until we actually have a builder that's
configured to run fint, at which point I may discover some bugs in this
logic.

I also deleted a couple variant-specific test cases from
build/examples/full.py that didn't provide any necessary test coverage.

Original-Bug: 60997
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/436274
Original-Revision: fc877895b00e7ed5571984d24f6be19b048f094c
GitOrigin-RevId: 90b40345be21927e7b622f993d27ac7daa6c31fe
Change-Id: I60adc510327837749fa63f9f777077754b16bb79