[roll] Roll fuchsia [brcmfmac] Add BRCMF_DBG_EVENT macro to make brcmf_event_msg logging consistent
There are a number of handlers used to process various events sent to
the driver from the firmware. This change makes event handler log the
contents of the event that triggered it to be logged in a consistent
manner. Note that logging each event from within each handler is
necessary because the reason code in each event message can depend on
the type of event.

Test: None necessary. Logging change only.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/434916
Original-Revision: 217576853ffe85321bf4dbc6e04f0e709d18410d
GitOrigin-RevId: 627efe1b175b22a4b0d1a0fae35a33898258329e
Change-Id: Ibfeba5490cce2a8e02b65e106aa028b97d9fc691