[roll] Roll fuchsia [flatland] Wire up buffer collection release fences.
When Flatland::DeregisterBufferCollection() was originally
implemented, deregistration with the Renderer was hooked up to a
release fence stored in the Flatland object in anticipation of
integration with the FrameScheduler. Now that Flatland accepts
release fences from clients and forwards them to the
FlatlandPresenter, having separate logic for this fence isn't
appropriate.

This change modifies Flatland to only make a special release fence
when there are actually buffers ready for release in a particular
Present(). This fence is appended to the to the list of client
provided release fences so it is signaled at the same time: when the
Present() has been fully processed.

Original-Fixed: 53330
TEST=fx test -o flatland_unittests
MULTIPLY: flatland_unittests (fuchsia): 30

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/435716
Original-Revision: 06f4b4ddca2c52dd564a754eb90e71435a7ffad6
GitOrigin-RevId: 06104925a72cd13f25617f354c947e1bda2111b7
Change-Id: I4b20c1881a261e87b1fcb2d69441fe4d64f188fc