[roll] Roll fuchsia [gidl][llcpp] Separate FIDL generation per GIDL file
Currently, the LLCPP benchmarks are built within the same
source set, with each benchmark importing a large file
with all of the FIDL headers. This follows what is done
for the rust benchmarks so a minimal number of FIDL files
is used for each GIDL file.

This does not have significant effects on build time, but
it does move towards a leaner build configuration and the
output filenames now match the specific GIDL test -- making
it easier to understand what is going on.

This is the LLCPP equivalent of a recent HLCPP change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/436554
Original-Revision: 7a25be6a652e4a049099f2d1eadfb3b9773c76f8
GitOrigin-RevId: 1798abedcb9892a946bb60652ade8519c725d316
Change-Id: I070c0a7e353d835c89ad52da3167d0c96c7f0f95