[roll] Roll fuchsia [timekeeper] Prepare to add HTTPSdate as a monitor source.
After this CL it should be a one line change to actually launch a
monitor, but the first time I actually tried launching httpsdate
was crashing.

This adds some basic logging for the monitor but will be much
more useful when per-time-source inpect is added in the next CL.

Note this also includes a couple of trait implementations (e.g.
Debug for a time source) that are needed in the next CL.

Original-Bug: 56858
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/435840
Original-Revision: 8a4339de802bae4ab5da323d8b7072b9b653dc19
GitOrigin-RevId: 1012e685f86936abf9833f11fe286925ccbed17d
Change-Id: Ia14e0135b0ebd8161006264424d0a0fb8a8a30d5