[roll] Roll fuchsia [storage][extractor] Add extractor library 3/n
The library in its current state provides a way to write an image
file. Storage software can use this library to dump the data that is
relevant to them.

The library is in rust. It provides C bindings(in c-bindings directory)
that can be used in C++ code base.

This patch adds image writer part of the library.

Test: Ran host side unit tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/432635
Original-Revision: 755e56260a5b3e3f9819ab76a8237b6737763c2e
GitOrigin-RevId: fc352661a5ffc2876483a34c47e795b2f3ed6eec
Change-Id: I885b50ad4a6579d52a8d9ed1ec7d186d65eea858