[roll] Roll fuchsia [fdf] Add a binding header target for vendors
This is part of a transition to an allowlist to prevent new uses of the
old bind rules. For other repositories (i.e. vendor repositories) we
need an escape hatch since we can't add them to the allowlist. This must
be set up first so that those repositories can transition before we set
up the allowlist.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/436757
Original-Revision: 5676457eb4a9fc8e79d17a7a8562442795eaaeb5
GitOrigin-RevId: 3fc6878240944253c9f6c08c89f8a181656ed851
Change-Id: Ia9a785438fc2654575b1e938293ed334f53764fa