[roll] Roll infra/recipes [auto_submit] Add recipe to process Auto-Submit.
Chromium uses logic through a Gerrit Frontend plugin called
"chromium-behavior" to handle Auto-Submit, however, for Fuchsia, we use
the "find-owners" plugin which constructs labels that are both `need`
and never satisfied which makes it difficult to determine whether a
change is becoming submittable when labels are set. This creates a
recipe that can run at some X minute frequency that finds all changes
that are both submittable* and Auto-Submit+1 and sets Commit-Queue+2 on
them.

This also updates the gerrit recipe_module to take a raw query string
for query_changes() instead of a dict as that surface did not allow
duplicate keys.

* Gerrit Submittable does not include unsubmitted parent or mergability.

Original-Bug: 10158
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/433538
Original-Revision: 6c556a737ec7639af5d014059b8bc549d61b837a
GitOrigin-RevId: 56217bb52e22188c6c7e2fc00a67e91584358771
Change-Id: Id043002f1312cb6a0f20cd0a4d4bf767e4b73e00