[roll] Roll infra/recipes [auto_roller] Skip extra 'roll cl status' step
Some Gerrit backend consistency issues mean that even after a
change-details request has indicated that a change has been merged, a
later request (even from the same host) may return an earlier non-merged
status. If this happens when we check that status after exiting the
"wait for CQ" loop, we'll mistakenly try to force-submit a change that's
already been merged (example: ci.chromium.org/b/8867068058843837024).

To avoid this sort of issue entirely, keep track of the last-seen status
returned by the Gerrit change-details endpoint so that we don't need to
make a follow-up "roll cl status" request to check the status after we
exit the "wait for CQ" loop. This way it's impossible for the
aforementioned sort of inconsistency to break the roller.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/436894
Original-Revision: 2c5ce26988d031c88bd21678542083fc1734d212
GitOrigin-RevId: f68dd3fe6f2d1e955ea02f721c7a76355a4c8f7e
Change-Id: Icaa758b9105487cb93dbbcc00ca9c7c3915972f6