[roll] Roll fuchsia [bt][rfcomm] Update Session to use closure for opened channels
Updates Session to used a channel opened callback, instead of directly
passing in the set of Clients. This clarifies the abstraction between an
RFCOMM Session, and the clients of the RfcommServer. Now, Session does
not need to know anything about its clients, and can simply deliver
opened RFCOMM channels to the provided closure.

This simplifies some of the tests in session.rs, and removes the
dependency on FIDL.

Original-Bug: 59582
Test: Updated bt-rfcomm-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/434814
Original-Revision: a7f7d85af15a74828e75d4194fa509289fd35775
GitOrigin-RevId: 9ea3640ac85de678b5de56f42febe3466e2cd3ee
Change-Id: Ieb74ed21f0f1bd5ec67f1284607fddd8f98ae5f1