[roll] Roll fuchsia [intl] Factors intl configuration out of basemgr
Removes the responsibility of serving fuchsia.intl.PropertyProvider
from basemgr.

This allows us to provide locale information for lower-level services
and brings the configuration of "low disk" devices closer to the
default.

Replaces the condition under which basemgr is started up - it seems
that on systems that take in basemgr.config there isn't a way to
start up basemgr automatically, so we do it this way.

Some background:

This is a 2nd attempt to merge, see revert at: http://fxr/436339.
The root cause for the revert was the change to `modular.config`, which
removed the only remaining stanza that forced `basemgr.cmx` to start
at system startup: there used to be more, but those apparently were
removed, accidentally leaving this one to be last.  In this edition
the change to `modular.config` was modified to use a different service
for eager startup of `basemgr.cmx`.

This change was then verified manually on emu, since no current
automated tests verify that basemgr is up and running.  In a separate
change I'm working on an E2E test that will guard basemgr startup
behavior.

Original-Bug: 42387
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/436240
Original-Revision: f4c6835578675297b26d4bef358559f4112d39eb
GitOrigin-RevId: 4411c38e0b9671ba71f3f5a9d95555e199744e6e
Change-Id: I39b1e9a237e02dcb3455e2d638c549e7c3a7db63