[roll] Roll fuchsia [fidl][llcpp] use a template for encode/decode
This doesn't change the behavior or the code size.

However, it makes modifying the encoder/decoder implementation easier.
We will be able to switch to an implementation which uses tables to an
implementation which uses generated methods without changing the user
code.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/436399
Original-Revision: bb20970fada89f6c98dc7cbe666ceee58814c990
GitOrigin-RevId: 0f277916454fbe743ff0c691e2cd52676c28119f
Change-Id: Id0293f853b981e65e4ae231e0704c48c761cdd94