[roll] Roll fuchsia [brcmfmac] Fix workqueue dedupe causing tx delays
When running iperf3 with UDP traffic it was observed that the rate would
drop to 0 bps intermittently or constantly. This was caused by the
workqueue deduplicating work which prevented the DPC from running in a
timely manner.

Specifically if the transmit queue was being emptied by a currently
running DPC and a new frame was being queued up right at the end of the
DPC then that new frame would not trigger the DPC to run again. Even
worse it would say that it triggered DPC even though it didn't. This
prevented future frames queueing from triggering the DPC because it was
already triggered. The only thing that would cause this to get unstuck
was an interrupt from firmware since this would run the DPC regardless.
The timing had to be such that the new frame is queued after the DPC has
completed but before the workqueue had updated the current work item. In
my tests this window is about 30 microseconds. In the test case frames
are queued up about every 100 microseconds so this issue was happening
fairly frequently as shown in the iperf3 test results.

It would be difficult to rectify this situation in the workqueue because
it would require some kind of cooperation between the workqueue and the
calling code to make sure the calling code knows when it's safe to queue
things up again.

This change modifies the workqueue to allow the same item to be
scheduled again if and only if it's currently running.  It will still be
deduped if it's waiting to be run. This may cause unnecessary DPC runs
every now and then but it seems like a fair tradeoff.

It seems that this case was made much more apparent because UDP was
used. With no acks coming back from the receiver (as would be the case
for TCP) the interrupt that would unlock the situation did not fire
nearly as frequently, making the problem much more visible. The longest
wait to unlock the stall was about 2-3 seconds in a handful of testruns.

BUG: 60770
BUG: 42151
Test: fx test brcmfmac-test
      fx test brcmfmac-sim-test
      iperf3 -c <x.x.x.x> -u -b 50M
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/436414
Original-Revision: 6c12d77832d68b2e4e2956a2a1680847312fc7e3
GitOrigin-RevId: 47345e2316994fd3190522a499d7be0bf922f5dc
Change-Id: I7317b79e45a54eca5d992ca65d9f077820c1356e