[roll] Roll fuchsia [scenic] Scenic Stressor v0
This CL proposes a stressor for scenic using OpaqueTest.

We can start scenic + root_presenter in an isolated component
manager, provide a fake display and spam input commands onto a
basic view (simple_app).

In v0, the stressor sends 50 taps to a 640x480 touchscreen
at the center of the screen for 5 seconds. In response, the display
should render new rectangle colors on each tap.

A rudimentary mock display driver was written in Rust to
drive the fake display.

This test is configured to run on AEMU environments only
because Vulkan is required for Scenic to work.

Test: fx test scenic-stress-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/429235
Original-Revision: 606b475afc827023c383f69fd4d806d19b5e0b97
GitOrigin-RevId: af91a57b7132a48323f04fa12b925469c86b74e4
Change-Id: I5a01e8c974641172f1b965a631b39f5c7bfdfcdc