[roll] Roll fuchsia [kernel][arm64][spinlock] minor tweak to the spinlock routine
In the inner spinlock loop, only loop back to the WFE if we detected
that the spinlock was already held by loading it and testing against
zero. If we fail to store because of an exclusive lock failure loop back
to the load instead of going through a WFE. Call this the two label
solution.

For all practical purposes the single label vs two label solution should
be the same but thinking about it the two label is a bit more logically
correct. Loop hard while you're failing a race with other cpus, WFE only
once you've lost.

Note there are conflicting examples that our spinlock is based on.

In the ARMv8 ARM version F.c in section K11.3.1 it describes a spinlock
implementation that matches ours with the exception of using 32bit
values instead of 64bit.

However, the ARM Trusted firmware at
https://github.com/ARM-software/arm-trusted-firmware/blob/master/lib/locks/exclusive/aarch64/spinlock.S#L50
at the time of writing uses the two label approach, which seems more
correct. Following this one as a golden example for now.

Original-Bug: 56818
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/435150
Original-Revision: faac0ee8cd08c015882e0bd990f1813191cd9bbc
GitOrigin-RevId: b1c41f7b4d2ea8639597052bec2021d9d2d4b68a
Change-Id: I0f7198d05ae1283de347d7dc3213770c43bc3f1f