[roll] Roll fuchsia [chromedriver] Add outputs field to target.
This will allow it to get copied to a path within the out dir. The
attached bug was caused because the path to the chromedriver started
with "../.." because it was relative to the out dir and that was causing
issues with the upload path when we were trying to upload it with the
artifacts through artifactory.

Original-Bug: 61695
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/437072
Original-Revision: fa31b368db674f1ab4c8843ee84a022c7fe4c9a3
GitOrigin-RevId: a8c42d4c123c68782f83787d199c8b321b4a7166
Change-Id: Ifd663b6902f63a2dd9a3f440daee520fd8c3a3ee