[roll] Roll fuchsia [ui-input] input-synthesis: move synthesizer tests to submodule
Move the existing synthesizer tests into a submodule, to prepare
for adding a set of new tests which exercise a different behavior
of the synthesizer.

This change doesn't affect the semantics of the existing tests.
It just helps us group things in the source code (as the number of
tests increase), and helps us more easily parse test output
(thanks to the grouping).

Original-Bug: 60789
Test: fx test input-sythensis-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/434929
Original-Revision: 6eb37ade320c914e468276798ad781e7bd83d05c
GitOrigin-RevId: 870eb173999f38303ed254c97dc537253d74d562
Change-Id: I7129f29b6f661d59cc7e33834d0ae30f32aef56f