[roll] Roll fuchsia [gidl][walker] Separate FIDL generation per GIDL file
Currently, the walker benchmarks are built within the same
source set, with each benchmark importing a large file
with all of the FIDL headers. This follows what is done
for the rust benchmarks so a minimal number of FIDL files
is used for each GIDL file.

This does not have significant effects on build time, but
it does move towards a leaner build configuration and the
output filenames now match the specific GIDL test -- making
it easier to understand what is going on.

This is the walker equivalent of a recent LLCPP change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/436628
Original-Revision: b833db6bf4015855c56ed1f8690c04b142a2a870
GitOrigin-RevId: f5f554d998d86a2086a06d55680224e035c5a14c
Change-Id: I897ccb5501d5d3f9bf3fa7018ff6b2cbc3929c91