[roll] Roll fuchsia [fidl][benchmarks] Remove gidl_multiple_file_source_set
This was a hacky solution to splitting up files. For reference
benchmarks, only a small number of GIDL files are currently needed so
they should be able too be built together. The reference benchmarks are
more difficult to split up in a clean way from the other benchmarks
because of the pairing between reference implementation and GIDL / FIDL
files that build the input value.

Original-Fixed: 52371

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/436632
Original-Revision: 18e435bb0826fd82ad1f8bae181160ab5ae46edb
GitOrigin-RevId: b30d60c5b233aa9d8cce5d71e50eb22e67075cc0
Change-Id: I770f4176cfa65c83d9778bacb5611f6c23f19c29