[roll] Roll infra/recipes [build] Simplify tool_attrib helper
The only use case for passing a lambda to `tool_attrib` was to select
the x64 Linux version of the tool to isolate, rather than the version of
the tool for the current platform.

We can simplify the code by adding a single `linux_x64` argument instead
of allowing a lambda. If `linux_x64` is set, we'll always isolate the
x64 Linux version of the tool.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/436636
Original-Revision: 628b16afe117cea3fa4ce661524d2c8b16fcfcfc
GitOrigin-RevId: 3ed888736782290f18f45b8283e9cb288c504083
Change-Id: Ibd6d96e166801d0909f092b022854def28751403