[roll] Roll fuchsia [garnet] Add fuchsia.media.Audio to SystemServices
Because fuchsia.media.Audio has been migrated to audio_core.cmx, it must
be allowed as a system-service for run-test-component environments,
such that it can be accessed through the system AudioCore.

Before, fuchsia.media.Audio was mapped as an injected-service through
audio.cmx. It can no longer be mapped as an injected-service, because we
do not want to inject another instance of audio_core.cmx along with the
existing AudioCore system-service, so we must allow fuchsia.media.Audio
to be a system-service.

Original-Bug: 57744

Test: no behavior change
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/436682
Original-Revision: e1adc58b8e1330b91846cb29723633f53545e7c0
GitOrigin-RevId: 52db1ef2c9c8fa6f04818bcabb94528701d16661
Change-Id: I5bce14ae50c48bac230938fa1e62e5c8869b4733