[roll] Roll fuchsia [SetUI][Storage] Refactor serialization.
SettingService structs are currently serialized by
device_storage and store_accessor. serde methods
are called directly on the types, which implement
Serialize and Deserialize.

This interaction limits the type's participation in
how it is serialized. This changelist moves this
behavior to be part of the DeviceStorageCompatible
trait with a default implementation matching the
current behavior. This will give flexibility in
the future for types to implement their own
serialization behavior.

Original-Bug: 61061
Test: fx test -o setui_service_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/436994
Original-Revision: 9be49020a7cbffb8b6f213e2852ac20516cf48e0
GitOrigin-RevId: a168389c9f21707eb5b970a1073d5b29970c32cb
Change-Id: If91be37f76c15e8d8208ca7c0e964dbcc78a0d01