[roll] Roll fuchsia [testsharder] apply safety factor to target duration when multiplying
When a multiplier's TotalRuns is not set, testsharder automatically
sets the number of runs to try to fill up targetDuration. If the test
takes much longer than its median (as recorded in the test durations
file), then the shard may time out. Users have seen this in the past
when using the MULTIPLY feature in their commit messages, and it's
quite annoying.

Once we automatically multiply tests (see attached bug), it will be
extremely important to avoid this failure mode, since users will not
be able to easily work around it (as they can today by specifying
total runs in the commit message).

Original-Bug: 50301
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/437296
Original-Revision: 4997fe108d180bcb628314ce9b78fc48a17b7051
GitOrigin-RevId: 62aec2ff2b386540f3659939f49172e040db8d83
Change-Id: I495c1662d4cad22d0a7368332dfc64db887349f1