[roll] Roll fuchsia [storage][extractor] Add extractor library 4/n
The library in its current state provides a way to write an image
file. Storage software can use this library to dump the data that is
relevant to them.

The library is in rust. It provides C bindings(in c-bindings directory)
that can be used in C++ code base.

This patch adds C bindings for image writer part of the library.

Test: Ran host side unit tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/432637
Original-Revision: 7f6d778845883976faf5a5a634faa5433b7fe3c1
GitOrigin-RevId: 52ab1508d695e7cd675a15d7c1b6efb5293d2733
Change-Id: Iabec9779c131fbc44a1b9814a184ccef11cbeca6