[roll] Roll infra/recipes [release] Add recipe to create release branches
This recipe is migrated from a downstream recipes repo. The
functionality is preserved with some miscellaneous cleanup of step
names, properties, tests, and documentation.

The primary motivation for moving the recipe to this repo is to make
breaking changes in the git recipe module easier to coordinate, by
avoiding the need for cross-repo non-trivial rolls.

A number of TODOs are left in the recipe to mark future cleanups with
the git and release recipe modules.

Original-Bug: 61767
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/437494
Original-Revision: f86ba4c85f6a473703bbd4dbcaa35e2c88765648
GitOrigin-RevId: 705687a189559b6bf7b7d50b8de8260ea1f75e73
Change-Id: I4f7007387bfd1ece80949c6f8c0365fd459296f7