[roll] Roll fuchsia [devices] Fix InitializeDriverHostSvcDir
A recent change reduced the set of allowed services that were forwarded
to a driver host. This set of services exluded the LogSink, which is
required to correctly forward driver logs to syslog.

This change adds fuchsia.logger.LogSink to the set of allowed services.

Original-Fixed: 61428
Original-Bug: 61773
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/437334
Original-Revision: c99e19124999d9ad3349ec32b28ee5956431e849
GitOrigin-RevId: 2a94c5b1a05f483fc7b580fc58f62e9a46d7443c
Change-Id: Ia9cd46f6a53b1870a2ca6a48631b659bb3f04a99