[roll] Roll fuchsia [fvm] Create Metadata wrapper class
Metadata is a wrapper class which hides the details about the way that
FVM metadata is laid out. It is intended to be the main way that
metadata is read/written, and provides convenience methods for accessing
the metadata structures.

An important difference with Metadata v.s. (e.g.) how the FVM driver
represents metadata is that Metadata stores *both* the A/B copies in
memory, and treats the entire buffer as an opaque blob of memory. This
is crucial for the later work on FVM snapshots, since eventually there
will be new metadata added which is in neither the existing A or B
copy (since the new metadata evolves independently). See fxbug.dev/59567
for details.

The downside of that is, of course, that we use twice as much memory to
store the metadata. This isn't too bad in practice, since the main scale
factor is the number of addressable slices, and for a typical
configuration (32K slices, 400M partition), that's only ~500K per
metadata copy.

Future work could reduce this memory cost by making the Metadata object
store a single copy of the metadata, but present a logical view of both
copies, where memory matters (primarily the driver).

Port two users of this class (fvm_host and
storage::volume_image::SparseImageReader) as a proof-of-concept. The
driver is *NOT* ported to use this yet, and will be ported later.

Also added an OWNERS file for zircon/tools/fvm.

Also fixed several bugs throughout fvm-host, such as:
- Not handling the last partition in the part tables

Original-Bug: 60709
Tested: New unit tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/432557
Original-Revision: dd617cb517e001991110a421380d8e0a9d1612ee
GitOrigin-RevId: 11eb03c4cf341df5300294870199d2626fe5b0dc
Change-Id: I8a52dff410d0d327573478b4ad76ab778782e35b