[roll] Roll infra/recipes [git] Simplify get_changed_files()
There doesn't seem to be any reason to use git diff-tree's `-z` flag
to return a null-separated list when getting the list of changed files.
It's more standard to use default newline separator.

Also add a `deleted` boolean parameter so that we can use this function
in the static_checks recipe instead of a raw `api.git()` call.

Accepting a list of mock changed files as a `test_data` argument is
slightly inconcsistent with most of the other git recipe module methods,
most of which accept a `step_test_data` lambda instead, but it is
convenient so I didn't see a need to change it.

Original-Bug: 60874
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/437902
Original-Revision: 1b3a7a570713b916b8c832e34ed41c943ed204c2
GitOrigin-RevId: a0314091896d2f427fec0282152adb334fa82de8
Change-Id: I79ce3568d867172aace0bfb322f505fd8edb26d2