[roll] Roll fuchsia [security][appmgr] Add system-updater allowlist
Add a new allowlist that will be used to restrict access to the
fuchsia.update.installer.Installer protocol to just those components
that should be able to initiate the installation of an update (and
tests).

A future change will use this allowlist in appmgr to enforce access.

Original-Bug: 61800
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/438177
Original-Revision: 761e434b55b897796081312ee6e3d8b7dca9a471
GitOrigin-RevId: 6e474d5fc77309cd9a60afd20deffeb572d69507
Change-Id: If36a915e2716d9930052a84266aa518fc4193ebe