[roll] Roll fuchsia [netdevice] Negotiate rx signalling threshold
The signalling threshold to provide the device with more buffers used to
be hard-coded to half of the FIFO depth, which may not be ideal for
devices that operate at different speeds or have different burst traffic
capabilities. This change provides a mechanism for device implementers
to tune that option.

Original-Fixed: 44835
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/433858
Original-Revision: 51bcc775166e1652488033fae43fb8e6c95c07d3
GitOrigin-RevId: 56f7d5cc0df9647f72f59ea3ec19bacd6a760a77
Change-Id: Ied2feedeac1ced5c1d29523ab2c04648060b9535