[roll] Roll fuchsia [scrutiny] Move rest service out of framework into app.
The Scrutiny command line is an interface to the framework and is part
of the application crate/binary. The rest network interface to Scrutiny
should also be in the app crate keeping the workings of the decoupled
agnostic from any particular interface.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/437200
Original-Revision: 2e5c6f53147b6634a48f496c579ae93f77e39e6d
GitOrigin-RevId: e886c8cbece161a7b05648292d4921eb3056b4f6
Change-Id: If3bf1ead8aacf3badfb3654e1e8f2db8c96e78eb