[roll] Roll fuchsia [fvm] Remove some FVM header helper functions.
PartitionTableOffset
PartitionTableByteSizeForUsablePartitions
AllocTableOffset
AllocTableLengthForDiskSize
MetadataSizeForDiskSize
BackupStart
SlicesStart

Renamed AllocTableLengthForUsableSliceCount to
AllocTableByteSizeForUsableSliceCount() for clarity.

Some of the computations in fvm-host-test.cc are simplified
dramatically. This is a result of inlining the implementation of
ComputeRequiredDataSize and then simplifying the math expression. Most
of the terms fall out. The code seemed to think it was testing much more
than it was in practice so there should actually be no difference in
coverage: it was just validating the file size matched the FvmContainer
size.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/438434
Original-Revision: d8df476fe885144453df5bec5e7b03ac2c10b927
GitOrigin-RevId: dec73703888b7e6b702078b71bd7cafc3ea7095d
Change-Id: Iac35b4e4e5f9add282029ed2f59433f75681f86f