[roll] Roll fuchsia [ui-input] input synthesis: make register_device a function
We want `register_device!()` to have different behavior depending
on which injection protocol(s) are present on the system, so
this functionality will be provided by a function in an `impl`
of the `Injector` trait.

As the first step towards that, transform `register_device!()`
from a macro into a function.

Original-Bug: 60789
Test: fx test input-synthesis-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/434931
Original-Revision: 6bcb62577f9d05bb7e9bdb2ca9558f124dfc7af9
GitOrigin-RevId: c67ab920681969d78a81b246efe46dd1d71add32
Change-Id: I9342648336f23cbeb88f6030f86ffa20e42a9b59