[roll] Roll fuchsia [sysmem] Work around leak during fuzzer destructor
On device unbind the channel error handler is run, which causes all the
VMOs to be released. However, since the loop is being torn down
TrackedParentVmo.zero_children_wait_ is canceled, which means that
TrackedParentVmo continues to hold a reference to the
LogicalBufferCollection and that causes a memory leak.

Tearing down sysmem doesn't happen on a production system, so we can
just add a workaround for the fuzzer. If we run the loop until idle
after closing the VMO and channel handles then everything should work
properly.

Original-Bug: 61392

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/438268
Original-Revision: 9076d1958a5287720c4ca5d7c718146315bdd5dd
GitOrigin-RevId: 285df2ecfa7f2151f632d0aa53d86661b2eab916
Change-Id: I6e6ff0d21f45f88ec4346cdacef410fb821b4d4f