[roll] Roll fuchsia [fidl] Temporarily disable fidlgen_hlcpp_test tests
clang-format is updated as a part of the clang roll and produces a
different output from what is expected in the golden file for
TestCodegenSource/union.test.json. To facilitate the roll, we will
temporarily disable this test, then turn it back on and regenerate the
tests.

Original-Bug: 61989
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/438974
Original-Revision: 549ebe3d69ec5ceca64a297127dc6d055da04e4e
GitOrigin-RevId: e6fe8c30ae29a848e3c5a3e03b7d5a4647a2dff6
Change-Id: I3ba70d273e7a44a895a9ddabed93a4f66c09c761