[roll] Roll fuchsia [fidl] Make WeakStubController reference counts thread safe
The HLCPP bindings are not thread safe. Ordinarily this is ok;
programmers are generally aware of this fact and ensure that FIDL
actions (including callbacks) are invoked on the appropriate thread.

However, the following illustrates a tricky corner case:

  // FIDL
  protocol Foo {
    SomeMethod() -> ();
  }

  // C++
  // SomeMethodCallback is a class (a functor) that contains
  // a WeakStubController member.
  void Foo::SomeMethod(SomeMethodCallback cb) {
    bg_thread.PostTask(std::move(cb));
  }

  void Foo::BackgroundThreadTask(SomeMethodCallback cb) {
    dowork();
    if (connection_closed()) {
      return;  // calls cb.~SomeMethodCallback()
    }
    fidl_thread.PostTask(cb);
  }

When the client invokes SomeMethod, the server needs to do work on a
background thread before it can respond to the FIDL call. Ordinarily,
the background thread will send the callback to the FIDL thread once
background work is complete. This ensures that all FIDL operations
happen on the FIDL thread.

However, if the FIDL connection is torn down before the work completes,
the background thread might delete the callback immediately without
first passing that callback to the FIDL thread. It turns out this is a
data race: the callback object internally contains a reference count (a
WeakStubController) which is not updated atomically. This reference
count might be shared by multiple callbacks. Therefore it is not legal
to delete the callback on the background thread -- it must be deleted on
the FIDL thread.

This is highly non-obvious. At the API level, the callback is just a
fit::function -- the SomeMethodCallback object is an unexposed
implementation detail. While fit::function documents that it may be
unsafe to invoke the callback from an arbitrary thread, it does not
document that the callback cannot be destructed from an arbitrary
thread. In particular, fit::function is move-only, and it's not obvious
that it would be unsafe to destruct a move-only object from a different
thread than the one it was created on.

I think this is a very easy mistake to make, and it's very easy to do
this is a corner case that is never covered by debug mode tests.

This CL removes that race by updating WeakStubController to use atomic
reference counts. This does _not_ make it safe to invoke the callback
from an arbitrary threads; it is only safe to destroy the callback from
an arbitrary threads.

Original-Fixed: 57079
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/437828
Original-Revision: ccbffeffd67092a31c669d4f771eb2d49ff7e5f4
GitOrigin-RevId: f0fba3f2027057c06191588e90dcf1dd5c064aca
Change-Id: I0aae5f68173aed7496ba361721b78471d085514d