[roll] Roll fuchsia [bt][a2dp] Start streams that are opened
In some cases, the peer will open a StreamEndpoint,
complete opening it, and then fail to start the stream.
They are not required to, so this is fine, but it means
the stream will sit indefinitely in an opened state.

After a timeout (currently 2 seconds) attempt to start the
stream instead if it was opened remotely.

Original-Bug: 50465
Test: added in bt-a2dp-unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/435076
Original-Revision: 13ac47e0d29a2d9c67ccc214156b2fd80320b2ae
GitOrigin-RevId: b21fc258c7ce765d1bbe2240b7521bb264baac5b
Change-Id: If7c72f4a974237d1761f0293e7b894ecb25f7815