[roll] Roll fuchsia [shell] Add error alternatives
The Alt combinator will discard a parse result if its second result
succeeds. This might be undesirable in the case of things like
Alt(Expression, Empty) because the failed parse of Expression might
contain useful error information, and if the parse is destined to fail,
that may be the point where we want it to fail.

Alt will now store the failed first parse result in the returned result,
if it successfully parsed more text than the succeeding parse. The Seq
combinator can then use this alternative return to effectively backtrack
if an error correction is needed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390514
Original-Revision: ba777c1bcccf9624c1da3672c5edf00e91b5e69a
GitOrigin-RevId: abd996a2c4efc3c469d43f149bdc1fdd74a8662e
Change-Id: I9d5702370dbba20604f033a47205edf144a4062a