[roll] Roll fuchsia [rust][fasync] Retry: Integrate instrumentation module into executor
[Retrying with timer-independent fix which removes non-determinism
from the test which previously flaked in some cases.]

Integrate the instrumentation module into the
fuchsia-async executor, so that stats are collected at runtime.

Add a method to the collector to generate a snapshot from
the current state of the collector.

Original-Bug: 58578
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/438360
Original-Revision: 2cbe0459e32456e5f17724a10e9d5212a58405c3
GitOrigin-RevId: 36c0ed327002cf0eda5aec26e3e534a01877a855
Change-Id: If70ac1124bf97fc1e5d9ef786529e101831d08db