[roll] Roll fuchsia [ninjatrace] Improve critical path visualization
Instead of rendering just the critical path in one thread, we instead
render it with all other actions and mark `critical_path` in their
categories, and elevate them to the top of the trace. This enables
highlighting all critical steps in the whole trace on chrome://tracing
for analysis.

Test: fx test --host -o ninjalog_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/439458
Original-Revision: 4d2ff5d797025cdada2022c1c1737fd889991f41
GitOrigin-RevId: 6694d81ee1636111fc02ab90aac5dca4de935f7b
Change-Id: I478d260d34d33996a792fb99e75556fe0eecd1e3